### 2020-01-21 #360 @richelbilderbeek
Submission: mcbette
Submitting Author: Richel Bilderbeek (@richelbilderbeek)
Repository: https://github.com/richelbilderbeek/mcbette
Version submitted: 1.8.2
Version accepted: TBD

---

```
Package: mcbette
Title: Model Comparison Using 'babette'
Version: 1.8.2
    person("Richel J.C.", "Bilderbeek", email = "richel@richelbilderbeek.nl", role = c("aut", "cre"), comment = c(ORCID = "0000-0003-1107-7049")))
Maintainer: Richel J.C. Bilderbeek <richel@richelbilderbeek.nl>
Description: 'BEAST2' (<http://www.beast2.org>) is a widely used
    Bayesian phylogenetic tool, that uses DNA/RNA/protein data
    and many model priors to create a posterior of jointly estimated
    phylogenies and parameters.
    'mcbette' allows to do a Bayesian model comparison over some
    site and clock models,
    using 'babette' (<https://www.github.com/ropensci/babette>).
    babette,
    beautier (>= 2.3),
    beastier (>= 2.0.25),
    mauricer,
    tracerer
    Rmpfr,
    testit
    ape,
    devtools,
    hunspell,
    lintr,
    nLTT,
    phangorn,
    rappdirs,
SystemRequirements: BEAST2 (http://www.beast2.org/)
```

`mcbette` automates a model comparison using the tool 'BEAST2' with one of its packages.

:warning: because `mcbette` is a wrapper for 'BEAST2' and because 'BEAST2' does not allow to do what `mcbette` does on Windows, this package does not work under the Windows operating system.

This package aims at scientists in the field of phylogenetics, as it allows to do a Bayesian model comparison of phylogenetic inference models. Or simpler: with `mcbette` you can pick the best model to build a phylogeny from a DNA sequence.

There is some overlap, as model comparison is an important topic in phylogenetics. Unique about `mcbette` is the way how it does so: it uses a novel 'BEAST2' package (called 'NS', shorthand of 'Nested Sampling') to directly estimate the marginal likelihood (a.k.a. the evidence) of an inference model. Because the technique of Nested Sampling is -in this field- novel, the 'BEAST2' package is novel and `mcbette` is novel.

I do intend to submit to JOSS, as it looks awesome! Because I did not expect this (for me, new) option, I am not prepared yet. I will await the acceptance of my submission first, than happily write this!

Not yet.
