### 2020-01-21 richelbilderbeek/mcbette#15 @richelbilderbeek
Add 'paper.md' for JOSS
See [my rOpenSci submission](https://github.com/ropensci/software-review/issues/360) in which I state I'd enjoy a JOSS article.
