### 2020-01-21 83d6f8f
Use correct spelling of name, progress #15
