### 2020-01-25 #360 @maelle
### Editor checks:

---

#### Editor comments

Thanks @richelbilderbeek for your submission! I'm now looking for reviewers.

In the meantime, I have a few comments/questions you can address:

* The link to BEAST2 in DESCRIPTION uses http, please replace it with https `<https://www.beast2.org/>` (I made a PR). If you used the http link in other packages, you can change it but that is obviously beyond the scope of my checks. :wink:

* Have you thought about providing guidance/examples on other CI services e.g. GH actions, Circle CI?

* In the README I think the existence of two modes (online, and offline on Mac and Linux) could be made clearer.

* I made a PR for replacing usethis with remotes in the installation guidelines, which you've merged.

* In the README section about installation, please write how many dependencies in total are needed by the package (I had to install >30 packages) and that rJava and Rmpfr are among them (on e.g. Ubuntu, it means installing system dependencies before installing the R packages),
to make it less of a surprise.

* In the FAQ, it says the other packages are not on CRAN, but they're on CRAN now, aren't they?

* Why is there no README.Rmd?

* Are your R packages listed somewhere on BEAST2 website?

`goodpractice` output

```r
  ✖ not use "Depends" in DESCRIPTION, as it can cause name clashes, and
    poor interaction with other packages. Use "Imports" instead.
  ✖ not import packages as a whole, as this can cause name clashes between
    the imported packages. Instead, import only the specific functions you need.
```

What are the reasons for importing the whole packages?

```r
  ✖ fix this R CMD check WARNING: LaTeX errors when creating PDF version.
    This typically indicates Rd problems. LaTeX errors found: ! LaTeX Error: File
    `ltxcmds.sty' not found. Type X to quit or <RETURN> to proceed, or enter new
    name. (Default extension: sty) ! Emergency stop. <read *> l.105
    \RequirePackage{ltxcmds}[2010/11/12] ^^M !  ==> Fatal error occurred, no output
    PDF file produced!
```

Maybe/probably just a setup thing?

---

Reviewers:
Due date:
