### 2020-01-25 #360 @maelle
You can now add a review badge [`rodev::use_review_badge(360)`](https://docs.ropensci.org/rodev/reference/use_review_badge.html).
