### 2020-01-26 #360 @maelle
I forgot a question: why are there expectations in examples? Can't these cases & expectations live in tests?
