### 2020-01-27 richelbilderbeek/mcbette#18 @richelbilderbeek
Use https in links to BEAST2 in all packages
From [here](https://github.com/ropensci/software-review/issues/360), thanks @maelle:

> The link to BEAST2 in DESCRIPTION uses http, please replace it with https <https://www.beast2.org/> (I made a PR). If you used the http link in other packages, you can change it but that is obviously beyond the scope of my checks. wink

Do this in all packages.
