### 2020-01-27 richelbilderbeek/mcbette#18 @maelle
http->https was part of the work we did on rOpenSci website btw https://ropensci.org/technotes/2019/12/19/urls-tidying/
