### 2020-01-27 richelbilderbeek/mcbette#19 @richelbilderbeek
Create mcbette_run
Currently, mcbette is both an R package as well as something that runs an R script.

I quote @maelle from [here](https://github.com/ropensci/software-review/issues/360):

 * In the README I think the existence of two modes (online, and offline on Mac and Linux) could be made clearer.

To resolve this, create a package called 'mcbette_run' that only runs a mcbette script.
