### 2020-01-27 richelbilderbeek/mcbette#20 @richelbilderbeek
Document dependencies more explicitly
I quote @maelle from [here](https://github.com/ropensci/software-review/issues/360):

> In the README section about installation, please write how many dependencies in total are needed by the package (I had to install >30 packages) and that rJava and Rmpfr are among them (on e.g. Ubuntu, it means installing system dependencies before installing the R packages), to make it less of a surprise.

Do so :+1:
