### 2020-01-27 richelbilderbeek/mcbette#21 @richelbilderbeek
babette is on CRAN
From [here](https://github.com/ropensci/software-review/issues/360) I quote @maelle:

> In the FAQ, it says the other packages are not on CRAN, but they're on CRAN now, aren't they?

Update the FAQ.
