### 2020-01-27 richelbilderbeek/mcbette#15 @richelbilderbeek
OK, I can guarantee a well-written draft this Saturday, but I cannot guarantee a polished article. Will that work for starters? Else, maybe I should postpone my review request?
