### 2020-01-27 richelbilderbeek/mcbette#15 @maelle
I suppose it's better to put the submission on hold indeed if @vbaliga is still available a bit later.
