### 2020-01-27 richelbilderbeek/mcbette#22 @richelbilderbeek
goodpractice is right
From [here](https://github.com/ropensci/software-review/issues/360) I quote goodpractice:

> goodpractice output

  ✖ not use "Depends" in DESCRIPTION, as it can cause name clashes, and
    poor interaction with other packages. Use "Imports" instead.
  ✖ not import packages as a whole, as this can cause name clashes between
    the imported packages. Instead, import only the specific functions you need.
