### 2020-01-27 richelbilderbeek/mcbette#23 @richelbilderbeek
Check setup
From [here](https://github.com/ropensci/software-review/issues/360), I quote:

  ✖ fix this R CMD check WARNING: LaTeX errors when creating PDF version.
    This typically indicates Rd problems. LaTeX errors found: ! LaTeX Error: File
    `ltxcmds.sty' not found. Type X to quit or <RETURN> to proceed, or enter new
    name. (Default extension: sty) ! Emergency stop. <read *> l.105
    \RequirePackage{ltxcmds}[2010/11/12] ^^M !  ==> Fatal error occurred, no output
    PDF file produced!

> Maybe/probably just a setup thing?

Fix it :+1:
