### 2020-01-27 richelbilderbeek/mcbette#24 @richelbilderbeek
Use README.Rmd?
I quote from [here](https://github.com/ropensci/software-review/issues/360):

> Why is there no README.Rmd?

Look into this.
