### 2020-01-27 #360 @richelbilderbeek
Thanks @maelle for the first wave of feedback :+1:

> The link to BEAST2 in DESCRIPTION uses http, please replace it with https <https://www.beast2.org/> (I made a PR). If you used the http link in other packages, you can change it but that is obviously beyond the scope of my checks. wink

 * [ ] Will do [here](https://github.com/richelbilderbeek/mcbette/issues/18)

> Have you thought about providing guidance/examples on other CI services e.g. GH actions, Circle CI?

I have thought about about it, but I felt it not worth the effort to put another CI in place. Usually I add an AppVeyor build to check on Windows, but because `mcbette` cannot work on Windows (due to BEAST2), I skipped the AppVeyor build (although I could add a trivial build).

You think I should?

 * In the README I think the existence of two modes (online, and offline on Mac and Linux) could be made clearer.

 *  [ ] Will do [here](https://github.com/richelbilderbeek/mcbette/issues/19)

 * I made a PR for replacing usethis with remotes in the installation guidelines, which you've merged.

Thanks!

 * In the README section about installation, please write how many dependencies in total are needed by the package (I had to install >30 packages) and that rJava and Rmpfr are among them (on e.g. Ubuntu, it means installing system dependencies before installing the R packages),
    to make it less of a surprise.

 * [ ] Will do so [here](https://github.com/richelbilderbeek/mcbette/issues/20)

> In the FAQ, it says the other packages are not on CRAN, but they're on CRAN now, aren't they?

They are, I will update the FAQ

 * [ ] Update FAQ [here](https://github.com/richelbilderbeek/mcbette/issues/21)

> Why is there no README.Rmd?

Because my other R packages on rOpenSci do not have it either. I'll look into this.

 * [ ] Look into this [here](https://github.com/richelbilderbeek/mcbette/issues/24)

> Are your R packages listed somewhere on BEAST2 website?

No. BEAST2 has its own (albeit BEAST2) packages. I think they are aware of babette. You think I should contact them?

>  I note the JOSS paper.md isn't there yet.

Correct, I made an Issue [here](https://github.com/richelbilderbeek/mcbette/issues/15). This JOSS item on rOpenSci is new to me, and I think it's cool. Because I am in the lost months of my PhD, though, I do need to be careful with my time. I suggest the deadline of ~~March 2nd~~ February 9th. Would that work?

> # goodpractice output

  ✖ not use "Depends" in DESCRIPTION, as it can cause name clashes, and
    poor interaction with other packages. Use "Imports" instead.
  ✖ not import packages as a whole, as this can cause name clashes between
    the imported packages. Instead, import only the specific functions you need.

> What are the reasons for importing the whole packages?

There are -really!- dozens of functions that need to be imported. OTOH, `goodpractice` is right!

 * [ ] Fix [here](https://github.com/richelbilderbeek/mcbette/issues/22)

  ✖ fix this R CMD check WARNING: LaTeX errors when creating PDF version.
    This typically indicates Rd problems. LaTeX errors found: ! LaTeX Error: File
    `ltxcmds.sty' not found. Type X to quit or <RETURN> to proceed, or enter new
    name. (Default extension: sty) ! Emergency stop. <read *> l.105
    \RequirePackage{ltxcmds}[2010/11/12] ^^M !  ==> Fatal error occurred, no output
    PDF file produced!

> Maybe/probably just a setup thing?

I'll check.

 * [ ] Fix [here](https://github.com/richelbilderbeek/mcbette/issues/23)
