### 2020-01-27 richelbilderbeek/mcbette#15 @richelbilderbeek
If you think so, I will follow your lead :+1:

I am sorry, the JOSS article is both new to me and seems great!
