### 2020-01-28 782187e
Fix https://github.com/richelbilderbeek/mcbette/issues/18
