### 2020-01-28 280eb56
Fix https://github.com/richelbilderbeek/mcbette/issues/18
