+++ README.md
@@ -9,161 +9,16 @@ Branch   |[![Travis CI logo](man/figures/TravisCI.png)](https://travis-ci.org)  
 
 Model Comparison using `babette`.
 
-## What is `mcbette`?
-
-`mcbette` allows you to do a Bayesian model comparison on a DNA alignment,
-in an easy-to-use way.
-
-To do so, mcbette uses the infrastructure of GitHub and Travis CI. For simple
-alignments, this costs nothing. Because everything happens online, 
-no installation is needed :+1:.
-
-You can 
- 
- * [watch a mcbette introduction video on YouTube](https://youtu.be/bLhrrSua8OM)
- * [download that video](http://richelbilderbeek.nl/mcbette.ogv)
-
-## How to use `mcbette` online?
-
-  1. Get a GitHub account
-  2. Get a Travis CI account (one can simply log in with GitHub)
-  3. Fork this repo
-  4. Activate Travis CI for your fork
-  5. Change the alignment FASTA file, called `my_alignment.fas`
-  6. See the results in the Travis CI build log
-
-## How to use `mcbette` offline?
-
-:warning: `mcbette` only works on Linux and Mac
 
 ### Installation
+
+:warning: `mcbette` only works on Linux and Mac
 
 ```r
 remotes::install_github("richelbilderbeek/mcbette", dependencies = TRUE)
 beastier::install_beast2()
 mauricer::install_beast2_pkg("NS")
 ```
-
-## Example
-
-`mcbette` starts from an alignment:
-
-![A DNA alignment](man/figures/alignment.png)
-
-`mcbette` estimates the evidence (also known as marginal likelihood)
-of each model.
-
-```{r}
-mcbette::est_marg_liks(
-  fasta_filename = system.file("extdata", "primates.fas", package = "mcbette")
-)
-```
-
-Here is the `mcbette` output of that alignment:
-
-```
-|site_model_name |clock_model_name   |tree_prior_name                | marg_log_lik| marg_log_lik_sd|    weight|
-|:---------------|:------------------|:------------------------------|------------:|---------------:|---------:|
-|GTR             |relaxed_log_normal |birth_death                    |    -1811.027|        6.080012| 0.0000000|
-|GTR             |relaxed_log_normal |coalescent_bayesian_skyline    |           NA|              NA| 0.0000000|
-|GTR             |relaxed_log_normal |coalescent_constant_population |    -1843.625|        9.135051| 0.0000000|
-|GTR             |relaxed_log_normal |coalescent_exp_population      |    -1793.983|        4.096881| 0.0302719|
-|GTR             |relaxed_log_normal |yule                           |    -1804.508|        5.838469| 0.0000008|
-|GTR             |strict             |birth_death                    |    -1798.271|        4.794403| 0.0004154|
-|GTR             |strict             |coalescent_bayesian_skyline    |           NA|              NA| 0.0000000|
-|GTR             |strict             |coalescent_constant_population |    -1803.812|        5.577658| 0.0000016|
-|GTR             |strict             |coalescent_exp_population      |    -1796.305|        5.341554| 0.0029673|
-|GTR             |strict             |yule                           |    -1798.523|        4.878357| 0.0003231|
-|HKY             |relaxed_log_normal |birth_death                    |    -1796.834|        4.383638| 0.0017480|
-|HKY             |relaxed_log_normal |coalescent_bayesian_skyline    |           NA|              NA| 0.0000000|
-|HKY             |relaxed_log_normal |coalescent_constant_population |    -1797.376|        4.821398| 0.0010168|
-|HKY             |relaxed_log_normal |coalescent_exp_population      |    -1805.887|        5.011170| 0.0000002|
[…]
