+++ DESCRIPTION
@@ -4,7 +4,7 @@ Version: 1.8.2
 Authors@R: c(
     person("Richèl J.C.", "Bilderbeek", email = "richel@richelbilderbeek.nl", role = c("aut", "cre"), comment = c(ORCID = "0000-0003-1107-7049")))
 Maintainer: Richèl J.C. Bilderbeek <richel@richelbilderbeek.nl>
-Description: 'BEAST2' (<http://www.beast2.org>) is a widely used
+Description: 'BEAST2' (<https://www.beast2.org>) is a widely used
     Bayesian phylogenetic tool, that uses DNA/RNA/protein data
     and many model priors to create a posterior of jointly estimated 
     phylogenies and parameters.
