+++ README.md
@@ -1,4 +1,4 @@
-# `mcbette`
+# `mcbette_run`
 
 Branch   |[![Travis CI logo](man/figures/TravisCI.png)](https://travis-ci.org)                                                                 |[![Codecov logo](man/figures/Codecov.png)](https://www.codecov.io)
 ---------|-------------------------------------------------------------------------------------------------------------------------------------|-------------------------------------------------------------------------------------------------------------------------------------------------------------------
@@ -12,7 +12,7 @@ Model Comparison using `babette`.
 
 ### Installation
 
-:warning: `mcbette` only works on Linux and Mac
+:warning: `mcbette_run` only works on Linux and Mac
 
 ```r
 remotes::install_github("richelbilderbeek/mcbette", dependencies = TRUE)
