### 2020-01-28 08cdcde
Travis simply only builds the package (moved run to 'mcbette_run' repo)
