### 2020-01-28 4f70908
Start fixing #25
