### 2020-01-28 7464598
Use downgrades MacOS version, from https://devops.stackexchange.com/a/8013
