+++ .travis.yml
@@ -18,6 +18,7 @@ matrix:
     os: linux
   - r: release
     os: osx
+    osx_image: xcode9.2
 
 r_github_packages:
   - jimhester/covr
