### 2020-01-28 7ecd290
Run goodpractice on Travis again
