### 2020-01-28 richelbilderbeek/mcbette#22 @richelbilderbeek
Added goodpractice to the build again, see [a Travis log here](https://travis-ci.org/richelbilderbeek/mcbette/jobs/642901977#L1211)
