### 2020-01-28 49e5025
Add rOpenSci badge, from https://github.com/ropensci/software-review/issues/360#issuecomment-578412340, thanks @maelle
