+++ scripts/issue_25_estimate_primates.R
@@ -14,14 +14,18 @@ beastier::save_nexus_as_fasta(
 
 df <- expand.grid(
   epsilon = c(1e20, 1, 1e-20),
-  particle_count = c(1, 16),
+  particle_count = c(1, 2),
   marg_log_lik = NA,
-  marg_log_lik_sd = NA
+  marg_log_lik_sd = NA,
+  ess = NA
 )
 
 for (i in seq_len(nrow(df))) {
   print(i)
-  inference_model <- create_test_ns_inference_model()
+  inference_model <- create_test_ns_inference_model(
+    site_model = create_hky_site_model(),
+    tree_prior = create_bd_tree_prior()
+  )
   inference_model$mcmc$epsilon <- df$epsilon[i]
   inference_model$mcmc$particle_count <- df$particle_count[i]
   inference_model$mcmc$chain_length <- 1e10
@@ -31,6 +35,7 @@ for (i in seq_len(nrow(df))) {
   )
   df$marg_log_lik <- evidence$marg_log_lik
   df$marg_log_lik_sd <- evidence$marg_log_lik_sd
+  df$ess <- evidence$ess
 }
 df
 
