### 2020-01-31 f8fe0ea
Progress JOSS paper, #15
