### 2020-01-31 9e1fc3b
Progress JOSS paper, #15
