+++ man/figures/marg_liks_joss.md
@@ -1,0 +1,6 @@
+|site_model_name |clock_model_name |tree_prior_name | marg_log_lik| marg_log_lik_sd|    weight|
+|:---------------|:----------------|:---------------|------------:|---------------:|---------:|
+|JC69            |strict           |yule            |    -179.9266|        2.427167| 0.2823000|
+|HKY             |strict           |yule            |    -182.2376|        1.992356| 0.0279950|
+|TN93            |strict           |yule            |    -179.0459|        2.446301| 0.6810869|
+|GTR             |strict           |yule            |    -183.4157|        2.636430| 0.0086181|
