+++ paper.md
@@ -19,43 +19,79 @@ bibliography: paper.bib
 
 # Summary
 
-All species on Earth are related to one another.
-They inherit their features from their ancestors
- 
-And all species on Earth use DNA to store the blueprint of their features,
-which they inherited from their ancestors.
-That DNA carries traces of a species'
-evolutionary history, as species that are closely related have more similar
-DNA. The field of phylogenetics tries to conclude the evolutionary
-history of species from their DNA sequences. To do so, the
-researcher needs to pick a model of how evolution works. The problem is,
-that there are many models to pick from.
+All life on Earth is related to one another.
+We, humans, inherit our characteristics from our parents, which they
+inherted from theirs. Going back in time long enough (say, between 4-7 million
+years), we'll arrive at the ancestors that gave rise to both humans and chimps.
+At that early time, the human-chimp-ancestor mixed their heritable
+features freely, as all members of that species could produce offspring with
+another of the opposite gender. Alas, this would not last, as the 
+human-chimp-ancestor evolved into to seperate species. From that moment on,
+newly evolved features in humans, would stay in humans only.
+
+![](man/figures/posterior_primates_consensus_joss.png)
+
+> The evolutionary relationship between primates,
+> going from the past (left) to the present (right).
+> It shows that humans and chimps are the two species that are closest related.
+> H = Human, C = Chimp, G = Gorilla, O = Orangutan, S = Siamang.
+
+The evolutionary history of species leaves its traces in DNA, the 
+heritable material that contains the blueprint of an organism's features.
+The DNA consists of a 4 letter alphabet, which encodes the features of an organism.
+Each letter of this alphabet is named after one of the four nucleotides, 
+that are the encoding part of a DNA molecule.
+The DNA we have in our cells are a mix between our mother's and father's DNA,
+with a few novel mutations added.
+Because -at least theoretically- any human woman can produce offspring with 
+any human man, the DNA of our species gets mixed and novel mutations can
+spread through our population. As a consequence,
+the DNA between humans is more simular, then when compared to the DNA of any 
+other species. 
+
+![](man/figures/alignment_joss.png)
+
+> Schematic figure of the DNA sequences of 4 hypothetical species.
+> Species are numbered 1 to 4.
+> Of each species, 40 nucleotides are known.
+> Cells with identical colors depict the same nucleotide.
+
+Nowadays, we know the DNA sequences of many species. With these genetic
+codes, we should be able to reconstruct the evolutionary history of all 
+species. An evolutionary history is commonly depicted by a tree-like
+figure, called a phylogeny, or simply 'tree'. The field of phylogenetics 
+aims to construct the 'best' phylogeny from a the genetic codes of species.
 
 ![](man/figures/alignment_joss.png)
 ![](man/figures/arrow.png)
 ![](man/figures/phylogeny_joss.png)
-Caption: phylogenetics in a nutshell: deriving the evolutionary history
-of species (at the right) from their DNA sequences (at the left)
 
-``mcbette`` is an R package that helps to pick the best model. conclude 
-the evolutionary history of species from their DNA sequences. 
-The 'best way', in this context, is that what is 'simple enough, but not 
-simpler'. 
+> phylogenetics in a nutshell: deriving the evolutionary history
+> of species (at the right) from their DNA sequences (at the left)
 
-``mcbette`` builds heavily upon the ``babette`` R package [@Bilderbeek:2018],
-that uses the tool 
[…]
