### 2020-02-02 edb6ce7
Combine figures for JOSS paper
