### 2020-02-02 1af78bd
Fix JOSS paper references
