+++ paper.bib
@@ -1,0 +1,31 @@
+@article{Bilderbeek:2018,
+  title={babette: BEAUti 2, BEAST2 and Tracer for R},
+  author={Bilderbeek, Richel JC and Etienne, Rampal S},
+  journal={bioRxiv},
+  pages={271866},
+  year={2018},
+  publisher={Cold Spring Harbor Laboratory}
+}
+
+@article{Bouckaert:2019,
+  title={BEAST 2.5: An advanced software platform for Bayesian evolutionary analysis},
+  author={Bouckaert, Remco and Vaughan, Timothy G and Barido-Sottani, Joelle and Duch{\^e}ne, Sebasti{\'a}n and Fourment, Mathieu and Gavryushkina, Alexandra and Heled, Joseph and Jones, Graham and K{\"u}hnert, Denise and De Maio, Nicola and others},
+  journal={PLoS computational biology},
+  volume={15},
+  number={4},
+  pages={e1006650},
+  year={2019},
+  publisher={Public Library of Science}
+}
+
+@article{Russel:2019,
+  title={Model selection and parameter inference in phylogenetics using nested sampling},
+  author={Russel, Patricio Maturana and Brewer, Brendon J and Klaere, Steffen and Bouckaert, Remco R},
+  journal={Systematic biology},
+  volume={68},
+  number={2},
+  pages={219--233},
+  year={2019},
+  publisher={Oxford University Press}
+}
+
