+++ README.md
@@ -9,7 +9,9 @@ Branch   |[![Travis CI logo](man/figures/TravisCI.png)](https://travis-ci.org)  
 
 ![](man/figures/mcbette_logo.png)
 
-Model Comparison using `babette`.
+Model Comparison using `babette` [Bilderbeek & Etienne, 2018]
+using the BEAST2 [Bouckaert et al., 2019] nested sampling package
+as described in [Russell et al., 2019].
 
 ### Installation
 
@@ -47,12 +49,7 @@ Package                                                                       | 
 
 ## References
 
-Article about `babette`:
-
  * Bilderbeek, Richel JC, and Rampal S. Etienne. "babette: BEAUti 2, BEAST 2 and Tracer for R." Methods in Ecology and Evolution (2018). https://doi.org/10.1111/2041-210X.13032
-
-Article about nested sampling:
-
+ * Bouckaert R., Vaughan T.G., Barido-Sottani J., Duchêne S., Fourment M., Gavryushkina A., et al. (2019) BEAST 2.5: An advanced software platform for Bayesian evolutionary analysis. PLoS computational biology, 15(4), e1006650.
  * Russel, Patricio Maturana, et al. "Model selection and parameter inference in phylogenetics using nested sampling." Systematic biology 68.2 (2019): 219-233.
-
 
