### 2020-02-02 richelbilderbeek/mcbette#15 @richelbilderbeek
@maelle: I've added the JOSS paper, at [the place where one would expect it](https://github.com/richelbilderbeek/mcbette/blob/master/paper.md).

Closing this Issue :+1:
