### 2020-02-02 #360 @richelbilderbeek
@maelle: the JOSS paper is ready for review, but I still need some time (say, another weekend) to fix the other points you've mentioned.
