### 2020-02-08 eac66ae
Update README and FAQ, fix #21:
