+++ doc/README.md
@@ -11,3 +11,4 @@ dev.off()
 fasta_filename <- system.file("extdata", "primates.fas", package = "mcbette")
 png("alignment.png"); ape::image.DNAbin(ape::read.FASTA(fasta_filename)); dev.off()
 ```
+
