### 2020-02-08 94e65dd
Document dependencies, fix #20
