+++ R/mcbette.R
@@ -35,5 +35,126 @@
 #' }
 #' @docType package
 #' @name mcbette
-#' @import beautier tracerer beastier mauricer babette
+#' @importFrom beautier create_alpha_param
+#' @importFrom beautier create_bd_tree_prior
+#' @importFrom beautier create_beast2_beast_xml
+#' @importFrom beautier create_beast2_input
+#' @importFrom beautier create_beast2_input_file
+#' @importFrom beautier create_beast2_input_file_from_model
+#' @importFrom beautier create_beast2_input_from_model
+#' @importFrom beautier create_beta_distr
+#' @importFrom beautier create_beta_param
+#' @importFrom beautier create_cbs_tree_prior
+#' @importFrom beautier create_ccp_tree_prior
+#' @importFrom beautier create_cep_tree_prior
+#' @importFrom beautier create_clock_model
+#' @importFrom beautier create_clock_model_from_name
+#' @importFrom beautier create_clock_model_rln
+#' @importFrom beautier create_clock_model_strict
+#' @importFrom beautier create_clock_models
+#' @importFrom beautier create_clock_models_from_names
+#' @importFrom beautier create_clock_rate_param
+#' @importFrom beautier create_data_xml
+#' @importFrom beautier create_distr
+#' @importFrom beautier create_distr_beta
+#' @importFrom beautier create_distr_exp
+#' @importFrom beautier create_distr_gamma
+#' @importFrom beautier create_distr_inv_gamma
+#' @importFrom beautier create_distr_laplace
+#' @importFrom beautier create_distr_log_normal
+#' @importFrom beautier create_distr_normal
+#' @importFrom beautier create_distr_one_div_x
+#' @importFrom beautier create_distr_poisson
+#' @importFrom beautier create_distr_uniform
+#' @importFrom beautier create_exp_distr
+#' @importFrom beautier create_gamma_distr
+#' @importFrom beautier create_gamma_site_model
+#' @importFrom beautier create_gtr_site_model
+#' @importFrom beautier create_hky_site_model
+#' @importFrom beautier create_inference_model
+#' @importFrom beautier create_inv_gamma_distr
+#' @importFrom beautier create_jc69_site_model
+#' @importFrom beautier create_kappa_1_param
+#' @importFrom beautier create_kappa_2_param
+#' @importFrom beautier create_lambda_param
+#' @importFrom beautier create_laplace_distr
+#' @importFrom beautier create_log_normal_distr
+#' @importFrom beautier create_loggers_xml
+#' @importFrom beautier create_m_param
+#' @importFrom beautier create_mcmc
+#' @importFrom beautier create_mcmc_nested_sampling
+#' @importFrom beautier create_mean_param
+#' @importFrom beautier create_mrca_prior
+#' @importFrom beautier create_mu_param
+#' @importFrom beautier create_normal_distr
+#' @importFrom beautier create_ns_mcmc
+#' @importFrom beautier create_one_div_x_distr
+#' @importFrom beautier create_param
+#' @importFrom beautier create_param_alpha
+#' @importFrom beautier create_param_beta
+#' @importFrom beautier create_param_clock_rate
+#' @importFrom beautier create_param_kappa_1
+#' @importFrom beautier create_param_kappa_2
+#' @importFrom beautier create_param_lambda
+#' @importFrom beautier create_param_m
+#' @importFrom beautier create_param_mean
+#' @importFrom beautier create_param_mu
+#' @importFrom beautier create_param_rate_ac
+#' @importFrom beautier create_param_rate_ag
+#' @importFrom beautier create_param_rate_at
+#' @importFrom beautier create_param_rate_cg
+#' @importFrom beautier create_param_rate_ct
[…]
