+++ vignettes/demo.Rmd
@@ -86,10 +86,10 @@ only differing in their nucleotide substitution model:
 
 ```{r}
 if (can_run_mcbette()) {
-  inference_model_1 <- create_test_ns_inference_model(
+  inference_model_1 <- beautier::create_test_ns_inference_model(
     site_model = beautier::create_jc69_site_model()   
   )
-  inference_model_2 <- create_test_ns_inference_model(
+  inference_model_2 <- beautier::create_test_ns_inference_model(
     site_model = beautier::create_gtr_site_model()   
   )
   inference_models <- c(list(inference_model_1), list(inference_model_2))
