+++ NAMESPACE
@@ -8,6 +8,7 @@ export(est_marg_lik)
 export(est_marg_liks)
 export(interpret_bayes_factor)
 importFrom(beastier,create_beast2_options)
+importFrom(beastier,is_beast2_installed)
 importFrom(beautier,create_alpha_param)
 importFrom(beautier,create_bd_tree_prior)
 importFrom(beautier,create_beast2_beast_xml)
@@ -129,3 +130,4 @@ importFrom(beautier,create_treelog_xml)
 importFrom(beautier,create_uniform_distr)
 importFrom(beautier,create_xml_declaration)
 importFrom(beautier,create_yule_tree_prior)
+importFrom(mauricer,is_beast2_ns_pkg_installed)
