### 2020-02-10 0ca8b60
Add 'mcbette_self_test', @Giappo will enjoy this, fix #26
