+++ R/mcbette.R
@@ -156,7 +156,13 @@
 #' @importFrom beautier create_uniform_distr
 #' @importFrom beautier create_xml_declaration
 #' @importFrom beautier create_yule_tree_prior
+#' @importFrom beastier check_beast2
 #' @importFrom beastier create_beast2_options
+#' @importFrom beastier install_beast2
 #' @importFrom beastier is_beast2_installed
+#' @importFrom beastier is_on_travis
+#' @importFrom beastier uninstall_beast2
+#' @importFrom mauricer install_beast2_pkg
 #' @importFrom mauricer is_beast2_ns_pkg_installed
+#' @importFrom mauricer uninstall_beast2_pkg
 NULL
