+++ R/check_beast2_ns_pkg.R
@@ -1,0 +1,14 @@
+#' Checks if the BEAST2 'NS' package is installed.
+#'
+#' Will \link{stop} if not
+#' @export
+check_beast2_ns_pkg <- function() {
+  beastier::check_beast2()
+  if (!is_beast2_ns_pkg_installed()) {
+    stop(
+      "BEAST2 'NS' package not installed. ",
+      "Tip: use mauricer::install_beast2_pkg(\"NS\")"
+    )
+  }
+  invisible(NULL)
+}
