+++ man/mcbette_self_test.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/mcbette_self_test.R
+\name{mcbette_self_test}
+\alias{mcbette_self_test}
+\title{Performs a minimal \link{mcbette} run}
+\usage{
+mcbette_self_test()
+}
+\description{
+Performs a minimal \link{mcbette} run
+}
