+++ NAMESPACE
@@ -2,13 +2,19 @@
 
 export(calc_weights)
 export(can_run_mcbette)
+export(check_beast2_ns_pkg)
 export(create_mcbette_beast2_options)
 export(create_ns_inference_model)
 export(est_marg_lik)
 export(est_marg_liks)
 export(interpret_bayes_factor)
+export(mcbette_self_test)
+importFrom(beastier,check_beast2)
 importFrom(beastier,create_beast2_options)
+importFrom(beastier,install_beast2)
 importFrom(beastier,is_beast2_installed)
+importFrom(beastier,is_on_travis)
+importFrom(beastier,uninstall_beast2)
 importFrom(beautier,create_alpha_param)
 importFrom(beautier,create_bd_tree_prior)
 importFrom(beautier,create_beast2_beast_xml)
@@ -130,4 +136,6 @@ importFrom(beautier,create_treelog_xml)
 importFrom(beautier,create_uniform_distr)
 importFrom(beautier,create_xml_declaration)
 importFrom(beautier,create_yule_tree_prior)
+importFrom(mauricer,install_beast2_pkg)
 importFrom(mauricer,is_beast2_ns_pkg_installed)
+importFrom(mauricer,uninstall_beast2_pkg)
