### 2020-02-10 d954d7a
Fix vignette
