+++ vignettes/demo.Rmd
@@ -39,19 +39,7 @@ library(mcbette)
 To use `mcbette`, BEAST2 and the BEAST2 `NS` package must be installed:
 
 ```{r}
-if (rappdirs::app_dir()$os == "win") {
-  message("'mcbette' can only run on Linux and MacOS")
-} else if (!is_beast2_installed()) {
-  message(
-    "BEAST2 must be installed. ",
-    "Tip: use 'beastier::install_beast2()'"
-  )
-} else if (!is_beast2_ns_pkg_installed()) {
-  message(
-    "The BEAST2 'NS' package must be installed. ",
-    "Tip: use 'mauricer::install_beast2_pkg(\"NS\")'"
-  )
-}
+mcbette_self_test()
 ```
 
 If you've just gotten a message that you need to install either
@@ -86,11 +74,11 @@ only differing in their nucleotide substitution model:
 
 ```{r}
 if (can_run_mcbette()) {
-  inference_model_1 <- create_test_ns_inference_model(
-    site_model = create_jc69_site_model()   
+  inference_model_1 <- beautier::create_test_ns_inference_model(
+    site_model = beautier::create_jc69_site_model()   
   )
-  inference_model_2 <- create_test_ns_inference_model(
-    site_model = create_gtr_site_model()   
+  inference_model_2 <- beautier::create_test_ns_inference_model(
+    site_model = beautier::create_gtr_site_model()   
   )
   inference_models <- c(list(inference_model_1), list(inference_model_2))
   
