+++ R/mcbette.R
@@ -37,7 +37,6 @@
 #' @name mcbette
 #' @importFrom beautier create_alpha_param
 #' @importFrom beautier create_bd_tree_prior
-#' @importFrom beautier create_beast2_beast_xml
 #' @importFrom beautier create_beast2_input
 #' @importFrom beautier create_beast2_input_file
 #' @importFrom beautier create_beast2_input_file_from_model
