+++ NAMESPACE
@@ -17,7 +17,6 @@ importFrom(beastier,is_on_travis)
 importFrom(beastier,uninstall_beast2)
 importFrom(beautier,create_alpha_param)
 importFrom(beautier,create_bd_tree_prior)
-importFrom(beautier,create_beast2_beast_xml)
 importFrom(beautier,create_beast2_input)
 importFrom(beautier,create_beast2_input_file)
 importFrom(beautier,create_beast2_input_file_from_model)
