+++ R/mcbette.R
@@ -53,7 +53,6 @@
 #' @importFrom beautier create_clock_models
 #' @importFrom beautier create_clock_models_from_names
 #' @importFrom beautier create_clock_rate_param
-#' @importFrom beautier create_data_xml
 #' @importFrom beautier create_distr
 #' @importFrom beautier create_distr_beta
 #' @importFrom beautier create_distr_exp
@@ -78,7 +77,6 @@
 #' @importFrom beautier create_lambda_param
 #' @importFrom beautier create_laplace_distr
 #' @importFrom beautier create_log_normal_distr
-#' @importFrom beautier create_loggers_xml
 #' @importFrom beautier create_m_param
 #' @importFrom beautier create_mcmc
 #' @importFrom beautier create_mcmc_nested_sampling
@@ -118,7 +116,6 @@
 #' @importFrom beautier create_s_param
 #' @importFrom beautier create_scale_param
 #' @importFrom beautier create_screenlog
-#' @importFrom beautier create_screenlog_xml
 #' @importFrom beautier create_sigma_param
 #' @importFrom beautier create_site_model
 #' @importFrom beautier create_site_model_from_name
@@ -141,7 +138,6 @@
 #' @importFrom beautier create_test_treelog
 #' @importFrom beautier create_tn93_site_model
 #' @importFrom beautier create_tracelog
-#' @importFrom beautier create_tracelog_xml
 #' @importFrom beautier create_trait_set_string
 #' @importFrom beautier create_tree_prior
 #' @importFrom beautier create_tree_prior_bd
@@ -151,7 +147,6 @@
 #' @importFrom beautier create_tree_prior_yule
 #' @importFrom beautier create_tree_priors
 #' @importFrom beautier create_treelog
-#' @importFrom beautier create_treelog_xml
 #' @importFrom beautier create_uniform_distr
 #' @importFrom beautier create_xml_declaration
 #' @importFrom beautier create_yule_tree_prior
