### 2020-02-10 9f03608
Use @importFrom, but these functions cannot be found in doc example :confused:
