### 2020-02-10 4424cbf
Thanks @lintr-bot
