+++ man/est_marg_lik.Rd
@@ -69,9 +69,9 @@ if (can_run_mcbette()) {
   expect_true("ess" \%in\% names(marg_lik))
 
   # with these data types ...
-  expect_true(is_one_double(marg_lik$marg_log_lik))
-  expect_true(is_one_double(marg_lik$marg_log_lik_sd))
-  expect_true(is_one_double(marg_lik$ess))
+  expect_true(beautier::is_one_double(marg_lik$marg_log_lik))
+  expect_true(beautier::is_one_double(marg_lik$marg_log_lik_sd))
+  expect_true(beautier::is_one_double(marg_lik$ess))
 
   # with these values
   expect_true(marg_lik$marg_log_lik < 0.0)
