+++ man/figures/marg_liks_primates_joss.md
@@ -1,6 +1,0 @@
-|site_model_name |clock_model_name |tree_prior_name | marg_log_lik| marg_log_lik_sd|    weight|
-|:---------------|:----------------|:---------------|------------:|---------------:|---------:|
-|JC69            |strict           |yule            |    -1940.060|        3.667696| 0.0000000|
-|HKY             |strict           |yule            |    -1806.506|        4.981120| 0.0000011|
-|TN93            |strict           |yule            |    -1792.776|        4.397807| 0.9999989|
-|GTR             |strict           |yule            |    -1809.543|        5.727602| 0.0000001|
