### 2020-02-10 8cd6efb
Sort .Rbuildignore
