### 2020-02-10 2f15a67
Use tryCatch on 'mcbette_self_test'
