+++ tests/testthat/test-mcbette_self_test.R
@@ -3,7 +3,7 @@ test_that("use", {
   if (is_on_travis()) {
     # Store the current state
     was_beast2_installed <- is_beast2_installed()
-    was_beast2_ns_installed <- FALSE
+    was_beast2_ns_pkg_installed <- FALSE
     if (was_beast2_installed) {
       was_beast2_ns_pkg_installed <- is_beast2_ns_pkg_installed()
     }
@@ -33,11 +33,11 @@ test_that("use", {
     )
 
     # Restore current state
-    if (was_beast2_ns_pkg_installed) {
-      install_beast2_pkg("NS")
-    }
     if (was_beast2_installed) {
       install_beast2()
+    }
+    if (was_beast2_ns_pkg_installed) {
+      install_beast2_pkg("NS")
     }
   }
 })
