+++ tests/testthat/test-mcbette_self_test.R
@@ -34,9 +34,11 @@ test_that("use", {
 
     # Restore current state
     if (was_beast2_installed) {
+      expect_false(is_beast2_installed())
       install_beast2()
     }
     if (was_beast2_ns_pkg_installed) {
+      expect_false(!is_beast2_ns_pkg_installed())
       install_beast2_pkg("NS")
     }
   }
