+++ tests/testthat/test-mcbette_self_test.R
@@ -38,7 +38,7 @@ test_that("use", {
       install_beast2()
     }
     if (was_beast2_ns_pkg_installed) {
-      expect_false(!is_beast2_ns_pkg_installed())
+      expect_false(is_beast2_ns_pkg_installed())
       install_beast2_pkg("NS")
     }
   }
