### 2020-02-11 2e46673
Improve and activate tests
