+++ tests/testthat/test-mcbette_self_test.R
@@ -42,5 +42,10 @@ test_that("use", {
       expect_false(is_beast2_ns_pkg_installed())
       install_beast2_pkg("NS")
     }
+    expect_equal(was_beast2_installed, is_beast2_installed())
+    expect_true(
+        !was_beast2_ns_installed ||
+        was_beast2_ns_installed == is_beast2_ns_pkg_installed()
+    )
   }
 })
