+++ tests/testthat/test-est_marg_lik.R
@@ -71,7 +71,9 @@ test_that("use, too few taxa for CBS", {
 
 test_that("abuse", {
 
-  skip("Not now")
+  if (!beastier::is_beast2_installed()) return()
+  if (!mauricer::is_beast2_ns_pkg_installed()) return()
+
   # fasta_filename
   expect_error(
     est_marg_lik(fasta_filename = "nonsense"),
