### 2020-02-11 41a067c
Follow goodpractice advice: move packages to Suggests
