### 2020-02-11 44cc9e1
Improve figure
