### 2020-02-11 4a7518b
Improve legend
