### 2020-02-12 40e62d2
Improve bug report template
