+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -23,16 +23,15 @@ A clear and concise description of what you expected to happen.
 **Screenshots**
 If applicable, add screenshots to help explain your problem.
 
-**Desktop (please complete the following information):**
- - OS: [e.g. iOS]
- - Browser [e.g. chrome, safari]
- - Version [e.g. 22]
+**Environment:**
+Show the results of running the following script:
 
-**Smartphone (please complete the following information):**
- - Device: [e.g. iPhone6]
- - OS: [e.g. iOS8.1]
- - Browser [e.g. stock browser, safari]
- - Version [e.g. 22]
+```r
+library(mcbette)
+sessionInfo()
+beastier::get_beast2_version()
+knitr::kable(mauricer::get_beast2_pkg_names())
+```
 
 **Additional context**
 Add any other context about the problem here.
