### 2020-02-12 4e04380
Improve custom issue template text
