### 2020-02-12 1b8f9fb
Improve Bug Issue Template [skip ci]
