+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -11,11 +11,11 @@ assignees: ''
 A clear and concise description of what the bug is.
 
 **To Reproduce**
-Steps to reproduce the behavior:
-1. Go to '...'
-2. Click on '....'
-3. Scroll down to '....'
-4. See error
+Script to reproduce the behavior:
+
+```r
+# Your R script here, without this comment :-)
+```
 
 **Expected behavior**
 A clear and concise description of what you expected to happen.
