+++ tests/testthat/test-mcbette_report.R
@@ -1,0 +1,3 @@
+test_that("use", {
+  expect_output(mcbette_report())
+})
