+++ NAMESPACE
@@ -8,6 +8,7 @@ export(create_ns_inference_model)
 export(est_marg_lik)
 export(est_marg_liks)
 export(interpret_bayes_factor)
+export(mcbette_report)
 export(mcbette_self_test)
 importFrom(beastier,check_beast2)
 importFrom(beastier,create_beast2_options)
