### 2020-02-12 8cf7b79
Clean lint, thanks @lintr-bot
