### 2020-02-12 275e124
Update NEWS [skip ci]
