### 2020-02-12 ea619dd
Allow testing on Windows
