### 2020-02-12 10114c9
Disable non-Windows test on Windows
