### 2020-02-12 a520334
Update NEWS [skip ci]
