### 2020-02-12 cf028aa
Show minimal example [skip ci]
