### 2020-02-12 104cac7
Update README from 'devtools::build_readme()'
