### 2020-02-12 0c328f3
Show error in marginal likelihood estimation in READMEs
