### 2020-02-12 ff72bfc
Update NEWS [skip ci]
