### 2020-02-12 richelbilderbeek/mcbette#22 @richelbilderbeek
This is weird:

 * Tests that fail when called by `goodpractice` [here](https://travis-ci.org/richelbilderbeek/mcbette/jobs/649390903#L1162), where the `R CMD check` worked fine [heer](https://travis-ci.org/richelbilderbeek/mcbette/jobs/649390903#L1096)
 * Only 55% of all tests is run [here](https://travis-ci.org/richelbilderbeek/mcbette/jobs/649390903#L1152)
