### 2020-02-12 #360 @richelbilderbeek
Just a quick update: done everything, except that I am struggling with [some goodpractice issues](https://github.com/richelbilderbeek/mcbette/issues/22#issuecomment-585164030).

 * I did add a AppVeyor/Windows build, just to be sure the package can build under Windows.
 * I added a `README.Rmd` as requested. It's a cool and useful feature!
 * JOSS paper is ready for review

Will post when having solved those last things. Then -AFAICS- the package is ready for review :+1:
