### 2020-02-12 8f461d7
Travis shows install statuses
