+++ .travis.yml
@@ -43,7 +43,9 @@ before_install:
 before_script:
   - # echo "mcbette should build with and without BEAST2 installed"
   - Rscript -e "beastier::install_beast2()"
+  - Rscript -e "beastier::is_beast2_installed()"
   - Rscript -e 'mauricer::install_beast2_pkg("NS")'
+  - Rscript -e 'mauricer::is_beast2_ns_pkg_installed()'
 
 script:
   - travis_wait 30 R CMD build .
