### 2020-02-12 richelbilderbeek/mcbette#22 @richelbilderbeek
Local run:

```
It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 90% of
    code lines are covered by test cases.

    R/check_beast2_ns_pkg.R:8:NA
    R/check_beast2_ns_pkg.R:9:NA
    R/check_beast2_ns_pkg.R:10:NA
    R/check_beast2_ns_pkg.R:11:NA
    R/est_marg_lik.R:79:NA
    ... and 16 more lines
```
