### 2020-02-12 richelbilderbeek/mcbette#22 @richelbilderbeek
Also, [covr::codecov failed](https://travis-ci.org/richelbilderbeek/mcbette/jobs/649410445#L1122):

```
$ if [ "$TRAVIS_OS_NAME" == "linux" ] && [ "$TRAVIS_R_VERSION_STRING" == "devel" ]; then travis_wait 20 Rscript -e 'covr::codecov()'; fi
Still running (1 of 20): Rscript -e covr::codecov()
The command Rscript -e covr::codecov() exited with 1.
Log:
Error in curl::curl_fetch_memory(url, handle = handle) :
  Timeout was reached: [codecov.io] Connection timed out after 10001 milliseconds
Calls: <Anonymous> ... request_fetch -> request_fetch.write_memory -> <Anonymous>
```
