### 2020-02-12 65bd15c
Add showing default BEAST2 path, thank @Giappo
