### 2020-02-20 ccd1740
Improve 'mcbette_report', fix #28, fix #29
