### 2020-02-20 ddc5ed2
Better test 'est_marg_lik'
