### 2020-02-20 ae02528
Fix style, thanks @lintr-bot
