### 2020-02-20 0725e02
First check BEAST2, then BEAST2 NS package
