### 2020-02-20 59ec82c
Report 'can_run_mcbette'
