### 2020-02-20 5d0df3c
Fix layout in 'mcbette_report' [skip ci]
