+++ R/mcbette_report.R
@@ -3,7 +3,7 @@
 mcbette_report <- function() {
   kat <- function(x) cat(x, sep = "\n")
   kat("****************")
-  kat("* mcbette *")
+  kat("* mcbette      *")
   kat("****************")
   kat(paste0("Can run mcbette: ", mcbette::can_run_mcbette()))
   kat(paste0("OS: ", rappdirs::app_dir()$os))
