### 2020-02-20 bad2baa
Expose that mcbette modifies user environment on Travis
