### 2020-02-20 richelbilderbeek/mcbette#22 @richelbilderbeek
Ah, I predict the test scripts modify the user environment. I'll expose with [this commit](https://github.com/richelbilderbeek/mcbette/commit/bad2baa7a41c0a494fd22cf5a34585d85c0c2105).
