### 2020-02-20 b1e1d95
Travis uses global variables to verify mcbette modifies environment
