### 2020-02-20 091c66b
Travis gives error if mcbette tests modify the user environment
