### 2020-02-20 8a32d55
Travis gives error if mcbette tests modify the user environment
