### 2020-02-20 99c2687
Let testthat test for change in user environment
