### 2020-02-20 f32f273
Thanks @lintr-bot
