+++ tests/testthat.R
@@ -2,17 +2,17 @@ library(testthat)
 library(mcbette)
 
 is_beast2_installed_before <- beastier::is_beast2_installed()
-is_beast2_ns_pkg_installed_before <- NA
+is_beast2_ns_pkg_installed_before <- NA # nolint long variable name
 if (is_beast2_installed_before) {
-  is_beast2_ns_pkg_installed_before <- mauricer::is_beast2_ns_pkg_installed()
+  is_beast2_ns_pkg_installed_before <- mauricer::is_beast2_ns_pkg_installed() # nolint long variable name
 }
 
 test_check("mcbette")
 
 is_beast2_installed_after <- beastier::is_beast2_installed()
-is_beast2_ns_pkg_installed_after <- NA
+is_beast2_ns_pkg_installed_after <- NA # nolint long variable name
 if (is_beast2_installed_before) {
-  is_beast2_ns_pkg_installed_after <- mauricer::is_beast2_ns_pkg_installed()
+  is_beast2_ns_pkg_installed_after <- mauricer::is_beast2_ns_pkg_installed() # nolint long variable name
 }
 
 expect_equal(
