### 2020-02-20 cbfd6c8
Add 'get_mcbette_state'
