+++ tests/testthat/test-get_mcbette_state.R
@@ -1,0 +1,3 @@
+test_that("use", {
+  expect_silent(get_mcbette_state())
+})
