+++ NAMESPACE
@@ -7,6 +7,7 @@ export(create_mcbette_beast2_options)
 export(create_ns_inference_model)
 export(est_marg_lik)
 export(est_marg_liks)
+export(get_mcbette_state)
 export(interpret_bayes_factor)
 export(mcbette_report)
 export(mcbette_self_test)
