+++ R/default_params_doc.R
@@ -23,6 +23,21 @@
 #'   as can be created by \link[beautier]{create_inference_model}
 #' @param inference_models a list of one or more inference models,
 #'   as can be created by \link[beautier]{create_inference_model}
+#' @param mcbette_state the \link{mcbette} state,
+#' which is a \link{list} with the following elements:
+#' \itemize{
+#'   \item{
+#'     beast2_installed
+#'       \link{TRUE} if BEAST2 is installed,
+#'       \link{FALSE} otherwise
+#'   }
+#'   \item{
+#'      ns_installed
+#'       \link{NA} if BEAST2 is not installed.
+#'       \link{TRUE} if the BEAST2 NS package is installed
+#'       \link{FALSE} if the BEAST2 NS package is not installed
+#'   }
+#' }
 #' @param mcmc an MCMC for the Nested Sampling run,
 #'   as can be created by \link[beautier]{create_mcmc_nested_sampling}
 #' @param os name of the operating system,
@@ -52,6 +67,7 @@ default_params_doc <- function(
   fasta_filename,
   inference_model,
   inference_models,
+  mcbette_state,
   mcmc,
   os,
   rng_seed,
