### 2020-02-20 1b28552
Add 'set_mcbette_state'
