+++ man/set_mcbette_state.Rd
@@ -1,0 +1,28 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/set_mcbette_state.R
+\name{set_mcbette_state}
+\alias{set_mcbette_state}
+\title{Set the \link{mcbette} state.}
+\usage{
+set_mcbette_state(mcbette_state)
+}
+\arguments{
+\item{mcbette_state}{the \link{mcbette} state,
+which is a \link{list} with the following elements:
+\itemize{
+  \item{
+    beast2_installed
+      \link{TRUE} if BEAST2 is installed,
+      \link{FALSE} otherwise
+  }
+  \item{
+     ns_installed
+      \link{NA} if BEAST2 is not installed.
+      \link{TRUE} if the BEAST2 NS package is installed
+      \link{FALSE} if the BEAST2 NS package is not installed
+  }
+}}
+}
+\description{
+Set the \link{mcbette} state.
+}
