+++ NAMESPACE
@@ -12,6 +12,7 @@ export(get_mcbette_state)
 export(interpret_bayes_factor)
 export(mcbette_report)
 export(mcbette_self_test)
+export(set_mcbette_state)
 importFrom(beastier,check_beast2)
 importFrom(beastier,create_beast2_options)
 importFrom(beastier,install_beast2)
