### 2020-02-20 953bd11
Check mcbette state transitions on Travis only
