### 2020-02-20 bf499e0
Check mcbette state in 'get_mcbette_state'
