+++ R/get_mcbette_state.R
@@ -34,5 +34,6 @@ get_mcbette_state <- function() {
   {
     state$beast2_installed <- FALSE
   }
+  testthat::expect_silent(check_mcbette_state(state))
   state
 }
