### 2020-02-20 727e045
Use 'set_mcbette_state' in tests that change the mcbette state
