+++ tests/testthat/test-set_mcbette_state.R
@@ -20,18 +20,18 @@ test_that("all transitions", {
   #
   # As a list:
   #
-  # Transition | From        | To
-  # -----------|-------------|-------------
-  # 1          | Neither     | BEAST2
-  # 2          | BEAST2      | BEAST2 + NA
-  # 3          | BEAST2 + NS | BEAST2
-  # 4          | BEAST2      | Neither
-  # 5          | Neither     | BEAST2 + NS
-  # 6          | BEAST2 + NS | Neither
+  # Transition | From        | To                                               # nolint this is no code
+  # -----------|-------------|-------------                                     # nolint this is no code
+  # 1          | Neither     | BEAST2                                           # nolint this is no code
+  # 2          | BEAST2      | BEAST2 + NA                                      # nolint this is no code
+  # 3          | BEAST2 + NS | BEAST2                                           # nolint this is no code
+  # 4          | BEAST2      | Neither                                          # nolint this is no code
+  # 5          | Neither     | BEAST2 + NS                                      # nolint this is no code
+  # 6          | BEAST2 + NS | Neither                                          # nolint this is no code
   #
   state_neither   <- list(beast2_installed = FALSE, ns_installed = FALSE)
-  state_beast2    <- list(beast2_installed = TRUE , ns_installed = NA)
-  state_beast2_ns <- list(beast2_installed = TRUE , ns_installed = TRUE)
+  state_beast2    <- list(beast2_installed = TRUE, ns_installed = NA)
+  state_beast2_ns <- list(beast2_installed = TRUE, ns_installed = TRUE)
   expect_silent(check_mcbette_state(state_neither))
   expect_silent(check_mcbette_state(state_beast2))
   expect_silent(check_mcbette_state(state_beast2_ns))
