### 2020-02-20 3fa3a0c
Check mcbette states' their logic values
