### 2020-02-20 0a4b51e
Cleaner mcbette state tests
