+++ tests/testthat/test-check_beast2_ns_pkg.R
@@ -30,5 +30,8 @@ test_that("use", {
 
     # Restore current state
     set_mcbette_state(before)
+
+    # Check if everything is the same again
+    expect_identical(before, get_mcbette_state())
   }
 })
