### 2020-02-20 359998b
Overcheck in 'set_mcbette_state'
