### 2020-02-20 20261f0
Uninstall NS if its desired state is FALSE or NA
