### 2020-02-20 077369c
Use testthat with info
