+++ .travis.yml
@@ -31,6 +31,9 @@ env:
     - beast2_ns_pkg_installed_after=-1
 
 r_github_packages:
+  - ropensci/beautier
+  - ropensci/beastier
+  - ropensci/babette
   - jimhester/covr
   - jimhester/lintr
   - MangoTheCat/goodpractice
