### 2020-02-23 richelbilderbeek/mcbette#22 @richelbilderbeek
Great, things work awesomly,

Yet, [the Travis CI logs](https://travis-ci.org/richelbilderbeek/mcbette/jobs/654023676#L1272) show:

```
── GP mcbette ──────────────────────────────────────────────────────────────────

It is good practice to

  ✖ checking tests ... Running ‘testthat.R’ ERROR Running the tests in
    ‘tests/testthat.R’ failed. Last 13 lines of output: ── 1. Error:
    Package style (@test-style.R#2) ──────────────────────────────────
    invalid 'path' argument Backtrace: 1. lintr::expect_lint_free() 2.
    lintr::lint_package(...) 3. base::normalizePath(path, mustWork =
    FALSE) 4. base::path.expand(path) ══ testthat results
    ═══════════════════════════════════════════════════════════ [ OK:
    194 | SKIPPED: 0 | WARNINGS: 0 | FAILED: 1 ] 1. Error: Package
    style (@test-style.R#2) Error: testthat unit tests failed Execution
    halted
```

I believe this is noise, as `mcbette` already tests for style.

OTOH, let's check...
