### 2020-02-23 richelbilderbeek/mcbette#22 @richelbilderbeek
No, `lintr` somehow loses track of the right path (pun unintented) :zap:
