### 2020-02-23 70a5c8f
Different order of testing for input
