### 2020-02-23 c35965e
Is CRAN version still good enough?
