+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: mcbette
 Title: Model Comparison Using 'babette'
-Version: 1.8.2
+Version: 1.8.3
 Authors@R: c(
     person("Richèl J.C.", "Bilderbeek", email = "richel@richelbilderbeek.nl", role = c("aut", "cre"), comment = c(ORCID = "0000-0003-1107-7049")))
 Maintainer: Richèl J.C. Bilderbeek <richel@richelbilderbeek.nl>
