### 2020-02-25 #360 @richelbilderbeek
Still to do:

 * [ ] contact BEAST2 developers
 * [ ] fix [noise in Travis logs](https://travis-ci.org/richelbilderbeek/mcbette/jobs/654811595#L1186) generated by goodpractice
