### 2020-03-06 76654dd
Use explicit 'stringsAsFactors = TRUE' in 'est_marg_liks'
