### 2020-03-08 #360 @richelbilderbeek
As far as I can see, the review is ready to roll :shipit: !

Regarding the goodpractice Issue, I've posted [a bug report at the goodpractice repo](https://github.com/MangoTheCat/goodpractice/issues/139). I hope this will allow this review to get started :rocket:
