### 2020-03-09 #360 @maelle
Thanks @richelbilderbeek!

@vbaliga @bjoelle I've set the review deadline to 2020-03-30. As a reminder, here's the [reviewer guide](https://devguide.ropensci.org/reviewerguide.html).
