### 2020-03-17 #360 @bjoelle
Installing `mcbette` isn't working for me due to issues with `rJava` (macOS 10.14.6).
The CRAN version of `rJava` is looking for a specific JDK which isn't the one on my machine. Compiling the package from source (following instructions here [https://zhiyzuo.github.io/installation-rJava/]) appeared to work, but running any function crashes the R console.
