### 2020-03-18 #360 @richelbilderbeek
I have observed that getting `rJava` to work is the hardest nut to crack. `mcbette` does build on MacOS on Travis CI, but well, that ain't your local computer :+1:
