### 2020-03-18 #360 @bjoelle
@richelbilderbeek I got it working on the cluster on Linux, so I will be able to review `mcbette` after all
