### 2020-03-19 richelbilderbeek/mcbette#30 @richelbilderbeek
Process feedback rOpenSci review
Request is [here](https://github.com/ropensci/software-review/issues/360).
