### 2020-03-20 #360 @bjoelle
>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

Estimated hours spent reviewing: 3h for installing + 4h for reviewing

---

This package is well written overall, and makes good use of one of the strengths of `babette`, which is to generate quickly multiple variations of an analysis.
I believe the package installation can be made easier, as detailed below. Otherwise, most of my comments concern the documentation, which I think is too sparse at the moment.

I had trouble with the package installation due to issues with `rJava`, and never got it working on my Mac (macOS 10.14.6). The CRAN version of `rJava` is looking for a specific JDK which isn't the one on my machine. Compiling it from source appeared to work, but running any function crashes the R console. Installing on a Linux cluster worked.

It looks to me as if the dependency on `rJava` is only used in the `get_default_java_path` function, which seems disproportionate with the amount of trouble it is. I would recommend exploring whether the dependency can be removed, and if not providing a work-around for people who cannot get `rJava` working - one option would be to ask users to provide the path to Java manually.

Other comments

 * instructions on how to configure to use an existing BEAST2 installation would be helpful, if it is possible. On Mac `mcbette` is not looking for it in the default place it is installed to, so that could lead to many duplicate installations.

 * the documentation (paper, README, vignette) was generally unclear as to which models can be selected using NS - the only mention I found was in the DESCRIPTION, saying that some substitution and clock models are available. It would be useful to include this in the README along with a list of available models. One example showing clock model comparison, in the vignette for instance, would also be useful.
 * in general, the documentation assumes that people using `mcbette` will already be familiar with the other packages of the suite, particularly `beautier`. I think this expectation needs to be stated much more clearly in the README and vignette, and that a link to an appropriate introduction vignette should be provided.
 * the example in the README should describe the tested model briefly (JC69, strict clock, etc).
 * there is a `create_ns_inference_model` function in the package, but the examples (README, vignette, package doc) use the `beautier::create_test_ns_inference_model` function instead - it's unclear what the difference is between the two, or which one should be used for a real analysis. If the test function is only used to set up a shorter analysis (as the README seems to imply ?), then it would be better to set the shorter analysis with the real function and modified parameters, and explain what is being done.
 * in the default parameter documentation, `beast2_options` should link to the `create_mcbette_beast2_options` function used in all the examples, instead of the  `beastier::create_beast2_options` function. Similarly `inference_model` should link to either `create_ns_inference_model` or `beautier::create_test_ns_inference_model` rather than `beautier::create_inference_model`.
 * when autocompleting in RStudio, the description shown on the function is the beginning of the Description section. This means that for instance `check_mcbette_state` shows only `Will stop otherwise.`. Same issue with `check_beast2_ns_pkg`, `interpret_bayes_factor`.
 * contrary to what the doc says, `est_marg_lik` does not return either `estimates` or `trees`.
 * the error message when `est_marg_lik` fails prints the inference model nicely, but not the BEAST2 options (it shows `beast2_1c6e311fc10d3.xml.stateNANAFALSETRUE`).
 * it would be useful to have an example in the `interpret_bayes_factor` doc of how to use that function with the output from `est_marg_liks`.

Minor comments

 * the doc for `est_marg_lik` links to itself, same with `est_marg_liks`
 * `est_marg_lik.R`, l16 "Actual" -> "Current"

 * the main feature I think is missing from the package at the moment is the ability to keep the log and tree output of the `babette` runs, in addition to the marginal likelihood. The documentation says that the trees are returned, but that does not appear to be the case. I would expect that users doing model selection will also be interested in the parameter estimates for the best model.
 * my other suggestion is to have a function similar to `interpret_bayes_factor` which could directly process the output from `est_marg_liks`, and print the table with the best model highlighted and how all the others compare to that one.

#### JOSS paper comments

 * two references are missing DOIs (required according to JOSS guidelines)
 * there is a `> Example text below` left in l18
 * the single quotes in the title (`'babette'`) give me a yaml parser error when building the file
 * l59-61 I think you need to clarify that you are only talking about living species here rather than all species. It's also unclear to me why you say we "should be able to reconstruct" rather than "can reconstruct".
 * l61-62 I would introduce phylogenies earlier, when showing the primates tree.
 * l70-71 I think this sentence is too vague. I would suggest replacing it with "However, constructing a phylogeny is non-trivial, as it requires the researcher to choose how to model the evolutionary process, including..."
 * l80-85 This section needs to mention that the 'best' model is always defined in regards to a particular dataset. It's explained later but needs to be made clear earlier.
 * l85 I don't think you can reuse this definition of 'best' for phylogenies - what would it mean for a phylogeny to be "simple" ?

Minor comments

 * l49 then -> than
 * l63 a the -> the
