+++ .travis.yml
@@ -18,7 +18,7 @@ matrix:
     os: linux
   - r: release
     os: osx
-    osx_image: xcode9.2
+    #osx_image: xcode9.2
 
 # It may be that this package uninstall BEAST2 and/or the packages
 # before and after the tests.
@@ -39,6 +39,8 @@ r_github_packages:
   #- ropensci/babette
 
 before_install:
+  # From https://stackoverflow.com/a/20042195
+  - Rscript -e "update.packages(checkBuilt = TRUE, ask = FALSE)"
   # ImageMagick
   - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo add-apt-repository -y ppa:opencpu/imagemagick; fi
   - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo apt-get update; fi
