### 2020-05-24 c2e8f56
Specify the correct return types in the doc
