+++ R/est_marg_lik.R
@@ -6,8 +6,6 @@
 #'   \item \code{marg_log_lik}: estimated marginal (natural) log likelihood
 #'   \item \code{marg_log_lik_sd}: estimated error of \code{marg_log_lik}
 #'   \item \code{esses} the Effective Sample Size
-#'   \item \code{estimates} the estimated marginal (natural) log likelihoods
-#'   \item \code{trees} the jointly-estimated posterior trees
 #' }
 #' @seealso
 #' \itemize{
