### 2020-05-24 da8c22f
Prove it is possible to call 'est_marg_liks' when BEAST2 is installed in a custom location
