### 2020-05-24 bd5b258
Document how to use BEAST2 when it is installed at a custom location
