### 2020-07-16 8fdd4eb
Thanks @lintr-bot
