+++ .travis.yml
@@ -18,7 +18,6 @@ matrix:
     os: linux
   - r: release
     os: osx
-    #osx_image: xcode9.2
 
 # It may be that this package uninstall BEAST2 and/or the packages
 # before and after the tests.
@@ -34,6 +33,10 @@ r_github_packages:
   - jimhester/covr
   - jimhester/lintr
   - MangoTheCat/goodpractice
+  - ropensci/beautier
+  - ropensci/tracerer
+  - ropensci/beastier
+  - ropensci/mauricer
 
 before_install:
   # From https://stackoverflow.com/a/20042195
