+++ DESCRIPTION
@@ -13,7 +13,7 @@ Description: 'BEAST2' (<https://www.beast2.org>) is a widely used
     using 'babette' (<https://www.github.com/ropensci/babette>).
 License: GPL-3
 LazyData: true
-RoxygenNote: 7.1.0
+RoxygenNote: 7.1.1
 VignetteBuilder: knitr
 URL: https://github.com/richelbilderbeek/mcbette
 BugReports: https://github.com/richelbilderbeek/mcbette/issues
