### 2020-07-16 d06fb40
Add rOpenSci reviewers, progress #30
