### 2020-07-16 c4b649d
Improve description of package goal, progress #30
