### 2020-07-16 c7cd227
Indeed, the JOSS short summary is missing :-|
