+++ paper.md
@@ -15,9 +15,11 @@ date: 21 January 2020
 bibliography: paper.bib
 ---
 
-> Example text below
+# Abstract
 
-# Summary
+TODO
+
+# Introduction
 
 All life on Earth is related to one another.
 We, humans, inherit our characteristics from our parents, which they
