+++ R/create_ns_inference_model.R
@@ -1,5 +1,6 @@
 #' Create an inference model to measure the evidence of.
 #'
+#' Create an inference model to measure the evidence of.
 #' To do so, the inference model is created as usual (see
 #' \link[beautier]{create_inference_model}), except
 #' for using a Nested Sampling MCMC (see \link[beautier]{create_ns_mcmc})
@@ -12,6 +13,7 @@
 #'   inference_model <- create_ns_inference_model()
 #'   expect_true(beautier::is_nested_sampling_mcmc(inference_model$mcmc))
 #' }
+#' @author Richèl J.C. Bilderbeek
 #' @export
 create_ns_inference_model <- function(
   site_model = beautier::create_jc69_site_model(),
