+++ R/interpret_bayes_factor.R
@@ -1,6 +1,6 @@
 #' Interpret a Bayes factor
 #'
-#' Interpretation from [1].
+#' Interpret a Bayes factor, using the interpretation from [1].
 #'
 #' \itemize{
 #'   \item [1] H. Jeffreys (1961). The Theory of Probability (3rd ed.).
@@ -16,6 +16,7 @@
 #' expect_equal(interpret_bayes_factor(12.5), "strong")
 #' expect_equal(interpret_bayes_factor(85.0), "very strong")
 #' expect_equal(interpret_bayes_factor(123.0), "decisive")
+#' @author Richèl J.C. Bilderbeek
 #' @export
 interpret_bayes_factor <- function(bayes_factor) {
 
