+++ man/can_run_mcbette.Rd
@@ -7,6 +7,7 @@
 can_run_mcbette()
 }
 \description{
+Can 'mcbette' run?
 Will return \link{TRUE} if:
 \itemize{
   \item (1) Running on Linux or MacOS
