### 2020-07-16 f17c560
Link to 'create_mcbette_beast2_options' wherever possible, progress #30
