+++ man/est_marg_liks.Rd
@@ -23,7 +23,7 @@ as can be created by \link[beautier]{create_inference_model}}
 
 \item{beast2_optionses}{list of one or more \code{beast2_options}
 structures,
-as can be created by \link[beastier]{create_beast2_options}.
+as can be created by \link{create_mcbette_beast2_options}.
 Use of reduplicated plural to achieve difference with
 \code{beast2_options}}
 
