### 2020-07-25 4f5b808
No testthat tests in examples, thanks Svetlana
