+++ R/get_mcbette_state.R
@@ -14,11 +14,7 @@
 #'   }
 #' }
 #' @examples
-#' library(testthat)
-#'
-#' state <- get_mcbette_state()
-#' expect_true("beast2_installed" %in% names(state))
-#' expect_true("ns_installed" %in% names(state))
+#' get_mcbette_state()
 #' @export
 get_mcbette_state <- function() {
   state <- list(
