+++ R/create_mcbette_beast2_options.R
@@ -6,13 +6,7 @@
 #' for model comparison) BEAST2 options structure,
 #' use  \link[beastier]{create_beast2_options}
 #' @examples
-#' library(testthat)
-#'
-#' beast2_options <- create_mcbette_beast2_options()
-#'
-#' # To use mcbette, the binary executable of BEAST2 must be used
-#' # (instead of the .jar file)
-#' expect_true(beastier::is_bin_path(beast2_options$beast2_path))
+#' create_mcbette_beast2_options()
 #' @author Richèl J.C. Bilderbeek
 #' @export
 create_mcbette_beast2_options <- function(
