+++ R/est_marg_liks.R
@@ -19,8 +19,6 @@
 #'     single inference mode
 #' }
 #' @examples
-#' library(testthat)
-#'
 #' if (can_run_mcbette()) {
 #'
 #'   # Use an example FASTA file
@@ -40,33 +38,11 @@
 #'   # Need as much beast2_optionses as inference models
 #'   beast2_optionses <- list(beast2_options, beast2_options)
 #'
-#'   df <- est_marg_liks(
+#'   est_marg_liks(
 #'     fasta_filename,
 #'     inference_models = inference_models,
 #'     beast2_optionses = beast2_optionses
 #'   )
-#'
-#'   # Testing the results
-#'   expect_true(is.data.frame(df))
-#'   expect_true("site_model_name" %in% colnames(df))
-#'   expect_true("clock_model_name" %in% colnames(df))
-#'   expect_true("tree_prior_name" %in% colnames(df))
-#'   expect_true("marg_log_lik" %in% colnames(df))
-#'   expect_true("marg_log_lik_sd" %in% colnames(df))
-#'   expect_true("weight" %in% colnames(df))
-#'   expect_true("ess" %in% colnames(df))
-#'
-#'   expect_true(is.factor(df$site_model_name))
-#'   expect_true(is.factor(df$clock_model_name))
-#'   expect_true(is.factor(df$tree_prior_name))
-#'   expect_true(!is.factor(df$marg_log_lik))
-#'   expect_true(!is.factor(df$marg_log_lik_sd))
-#'   expect_true(!is.factor(df$weight))
-#'
-#'   expect_true(all(df$weight >= 0.0))
-#'   expect_true(all(df$weight <= 1.0))
-#'   expect_true(sum(df$marg_log_lik < 0.0, na.rm = TRUE) > 0)
-#'   expect_true(sum(df$marg_log_lik_sd > 0.0, na.rm = TRUE) > 0)
 #' }
 #' @author Richèl J.C. Bilderbeek
 #' @export
