+++ man/interpret_bayes_factor.Rd
@@ -22,13 +22,7 @@ Interpret a Bayes factor, using the interpretation from [1].
 }
 }
 \examples{
-library(testthat)
-expect_equal(interpret_bayes_factor(0.5), "in favor of other model")
-expect_equal(interpret_bayes_factor(1.5), "barely worth mentioning")
-expect_equal(interpret_bayes_factor(8.5), "substantial")
-expect_equal(interpret_bayes_factor(12.5), "strong")
-expect_equal(interpret_bayes_factor(85.0), "very strong")
-expect_equal(interpret_bayes_factor(123.0), "decisive")
+interpret_bayes_factor(0.5)
 }
 \author{
 Richèl J.C. Bilderbeek
