+++ tests/testthat/test-est_marg_lik.R
@@ -163,11 +163,12 @@ test_that("use BEAST2 installed at a different location", {
   beastier::install_beast2(folder_name = folder_name)
   expect_true(beastier::is_beast2_installed(folder_name = folder_name))
 
-  bin_path <- beastier::get_default_beast2_bin_path(beast2_folder = folder_name)
-  expect_true(file.exists(bin_path))
+  beast2_bin_path <- beastier::get_default_beast2_bin_path(beast2_folder = folder_name)
+  expect_true(file.exists(beast2_bin_path))
   beast2_options <- create_mcbette_beast2_options(
     beast2_bin_path = beast2_bin_path
   )
+
   expect_silent(beastier::check_beast2_options(beast2_options))
 
   fasta_filename <- system.file("extdata", "simple.fas", package = "mcbette")
