### 2020-07-25 8208d95
Shorter lines, thanks @lintr-bot
