### 2020-07-26 ffd8c8d
Show available site, clock and tree models is README and vignette
