+++ README.md
@@ -16,10 +16,15 @@
 
 ![](man/figures/mcbette_logo.png)
 
-`mcbette` allows for doing a Model Comparison using `babette`
-\[Bilderbeek & Etienne, 2018\]. It does so using the BEAST2 \[Bouckaert
-et al., 2019\] nested sampling package as described in \[Russell et al.,
-2019\].
+`mcbette` allows for doing a Model Comparison using `babette` (hence the
+name), that is, given a (say, DNA) alignment, it compares multiple
+phylogenetic inference models to find the model that is likeliest to
+generate that alignment. With this, one can find the phylogenetic
+inference model that is simple enough, but not too simple.
+
+To do so, `mcbette` uses `babette` \[Bilderbeek and Etienne, 2018\] with
+the addition of using the BEAST2 \[Bouckaert et al., 2019\] nested
+sampling package as described in \[Russell et al., 2019\].
 
 ## Installation
 
@@ -43,29 +48,107 @@ mauricer::install_beast2_pkg("NS")
 
 ## Example
 
-This is a basic example which shows you how to solve a common problem:
-
 ``` r
 library(mcbette)
-
-# An example FASTA file
-fasta_filename <- system.file("extdata", "simple.fas", package = "mcbette")
-
-# A testing inference model with inaccurate (thus fast) marginal
-# likelihood estimation
-inference_model <- beautier::create_test_ns_inference_model()
-
-# Estimate the marginal likelihood
-marg_lik <- est_marg_lik(
-  fasta_filename = fasta_filename,
-  inference_model = inference_model
-)
-
-cat(marg_lik$marg_log_lik)
-#> -14.01602
-cat(marg_lik$marg_log_lik_sd)
-#> 1.78702
 ```
+
+Suppose we have a DNA alignment, obtained by sequencing multiple
+species:
+
+``` r
+fasta_filename <- beautier::get_beautier_path("anthus_aco_sub.fas")
+ape::image.DNAbin(ape::read.FASTA(fasta_filename))
+```
+
+<img src="man/figures/README-example_alignment-1.png" width="100%" />
+
+Note that this alignment holds too little information to really base a
+publishable research on.
+
+To create a posterior distribution of phylogenies from this alignment,
+one needs to specify an inference model. An inference model is (among
+others) a combination of a site model, clock model and tree model. See
+the ‘Available models’ section to see the available models.
+
+In this example we let two inference models compete.
+
+Here is the default inference model:
+
+``` r
+inference_model_1 <- create_ns_inference_model()
+print(
+  paste(
+    inference_model_1$site_model$name,
+    inference_model_1$clock_model$name,
+    inference_model_1$tree_prior$name
+  )
+)
+#> [1] "JC69 strict yule"
+```
+
+The competing model has a different site, clock and tree model:
+
+``` r
+inference_model_2 <- inference_model_1
+inference_model_2$site_model <- beautier::create_hky_site_model()
+inference_model_2$tree_prior <- beautier::create_bd_tree_prior()
+inference_model_2$clock_model <- beautier::create_rln_clock_model()
+```
+
+`mcbette` shows the evidence (also called marginal likelihood) for each
+inference model, which is the likelihood that a model has generated the
+data.
+
+``` r
+if (can_run_mcbette()) {
+  knitr::kable(
+    est_marg_liks(
+      fasta_filename = fasta_filename,
+      inference_models = list(inference_model_1, inference_model_2)
+    )
+  )
+}
+```
+
+| site\_model\_name | clock\_model\_name   | tree\_prior\_name | marg\_log\_lik | marg\_log\_lik\_sd | weight |      ess |
+| :---------------- | :------------------- | :---------------- | -------------: | -----------------: | -----: | -------: |
[…]
