### 2020-07-31 94b3d35
Rewrote paper to process the reviewers' comments generally, progress #30
