+++ paper.md
@@ -4,6 +4,9 @@ tags:
   - R
   - phylogenetics
   - model comparison
+  - nested sampling
+  - babette
+  - BEAST2
 authors:
   - name: Richèl J.C. Bilderbeek
     orcid: 0000-0003-1107-7049
@@ -11,9 +14,13 @@ authors:
 affiliations:
  - name: Theoretical & Evolutionary Community Ecology, TRES, GELIFES, University of Groningen
    index: 1
-date: 21 January 2020
+date: 31 July 2020
 bibliography: paper.bib
 ---
+
+![](man/figures/mcbette_logo.png)
+
+> ``mcbette`` logo
 
 # Abstract
 
@@ -24,105 +31,54 @@ evidence for having generated the alignment, from a set of models.
 In this way, the model that is 'simple enough, but not simpler' can
 be used to generate a phylogeny.
 
-# Introduction
-
-All life on Earth is related to one another.
-We, humans, inherit our characteristics from our parents, which they
-inherited from theirs. Going back in time long enough (say, between 4-7 million
-years), we'll arrive at the ancestors that gave rise to both humans and chimps.
-At that early time, the human-chimp-ancestor mixed their heritable
-features freely, as all members of that species could produce offspring with
-another of the opposite gender. Alas, this would not last, as the 
-human-chimp-ancestor evolved into to separate species. From that moment on,
-newly evolved features in humans, would stay in humans only.
-
-![](man/figures/posterior_primates_consensus_joss.png)
-
-> The evolutionary relationship between primates,
-> going from the past (left) to the present (right).
-> It shows that humans and chimps are the two species that are closest related.
-> Horizontal lines are not to scale.
-
-The evolutionary history of species leaves its traces in DNA, the 
-heritable material that contains the blueprint of an organism's features.
-The DNA consists of a 4 letter alphabet, which encodes the features of an organism.
-Each letter of this alphabet is named after one of the four nucleotides, 
-that are the encoding part of a DNA molecule.
-The DNA we have in our cells are a mix between our mother's and father's DNA,
-with a few novel mutations added.
-Because -at least theoretically- any human woman can produce offspring with 
-any human man, the DNA of our species gets mixed and novel mutations can
-spread through our population. As a consequence,
-the DNA between humans is more similar, than when compared to the DNA of any 
-other species. 
-
-![](man/figures/alignment_joss.png)
-
-> Schematic figure of a DNA alignment of 4 hypothetical species.
-> Species are numbered 1 to 4.
-> Of each species, 40 nucleotides are known.
-> Cells with identical colors depict the same nucleotide.
-
-Nowadays, we know the DNA sequences of many species. With these genetic
-codes, we are able to reconstruct the evolutionary history of all 
-species. An evolutionary history is commonly depicted by a tree-like
-figure, called a phylogeny, or simply 'tree'. The field of phylogenetics 
-aims to reconstruct the true phylogeny from the genetic codes of species.
-
-(that is, multiple
-DNA sequences arranged in such a way that most nucleotides
-are in the same relative position)
-
+# Statement of need 
 
 ![](man/figures/combined.png)
 
-> phylogenetics in a nutshell: deriving the evolutionary history
-> of species (at the right) from their DNA sequences (at the left)
+> Constructing a species phylogeny (at the right) 
+> from a DNA alignment (at the left)
+> using an evolutionary model (the arrow). 
+> ``mcbette`` allows to select an evolutionary model from a set of models
 
-Constructing a phylogeny from a DNA alignment, is non-trivial, 
[…]
