+++ paper.bib
@@ -18,6 +18,35 @@
   publisher={Public Library of Science}
 }
 
+@article{Drummond:2005,
+  title={Bayesian coalescent inference of past population dynamics from molecular sequences},
+  author={Drummond, Alexei J and Rambaut, Andrew and Shapiro, BETH and Pybus, Oliver G},
+  journal={Molecular Biology and Evolution},
+  volume={22},
+  number={5},
+  pages={1185--1192},
+  year={2005},
+  publisher={Society for Molecular Biology and Evolution}
+}
+
+@book{Drummond:2015,
+  title={Bayesian evolutionary analysis with {BEAST}},
+  author={Drummond, Alexei J and Bouckaert, Remco R},
+  year={2015},
+  publisher={Cambridge University Press}
+}
+
+@article{Nee:1994,
+  title={The reconstructed evolutionary process},
+  author={Nee, Sean and May, Robert M and Harvey, Paul H},
+  journal={Philosophical Transactions of the Royal Society of London B},
+  volume={344},
+  number={1309},
+  pages={305--311},
+  year={1994},
+  publisher={The Royal Society}
+}
+
 @article{Russel:2019,
   title={Model selection and parameter inference in phylogenetics using nested sampling},
   author={Russel, Patricio Maturana and Brewer, Brendon J and Klaere, Steffen and Bouckaert, Remco R},
