### 2020-07-31 #360 @richelbilderbeek
Nearly done processing all feedback. As expected: the package already got so much better :+1:
