### 2020-07-31 270736f
Progress in writing paper, progress #30
