### 2020-08-07 958766b
Fix style, thanks @lintr-bot
