+++ tests/testthat/test-create_ns_inference_model.R
@@ -1,4 +1,0 @@
-test_that("use", {
-  inference_model <- create_ns_inference_model()
-  expect_true(beautier::is_nested_sampling_mcmc(inference_model$mcmc))
-})
