+++ DESCRIPTION
@@ -36,7 +36,7 @@ URL: https://github.com/richelbilderbeek/mcbette
 BugReports: https://github.com/richelbilderbeek/mcbette/issues
 Imports:
     babette,
-    beautier (>= 2.3),
+    beautier (>= 2.4),
     beastier (>= 2.0.25),
     curl,
     mauricer,
