+++ NAMESPACE
@@ -5,7 +5,6 @@ export(can_run_mcbette)
 export(check_beast2_ns_pkg)
 export(check_mcbette_state)
 export(create_mcbette_beast2_options)
-export(create_ns_inference_model)
 export(est_marg_lik)
 export(est_marg_liks)
 export(get_mcbette_state)
