### 2020-08-25 3b06ae9
'create_ns_inference_model' is in beautier
