+++ R/est_marg_lik.R
@@ -57,7 +57,7 @@
 #' @export
 est_marg_lik <- function(
   fasta_filename,
-  inference_model = create_ns_inference_model(),
+  inference_model = beautier::create_ns_inference_model(),
   beast2_options = create_mcbette_beast2_options(),
   os = rappdirs::app_dir()$os
 ) {
