### 2020-08-25 7d17a17
Move 'create_mcbette_beast2_options' to 'beastier'
