+++ NAMESPACE
@@ -4,7 +4,6 @@ export(calc_weights)
 export(can_run_mcbette)
 export(check_beast2_ns_pkg)
 export(check_mcbette_state)
-export(create_mcbette_beast2_options)
 export(est_marg_lik)
 export(est_marg_liks)
 export(get_mcbette_state)
