### 2020-08-25 d972603
Moved 'create_mcbette_beast2_options' to 'beastier'
