+++ R/mcbette.R
@@ -2,7 +2,7 @@
 #'
 #' 'mcbette' does a model comparing using \link[babette]{babette},
 #' where the models are Bayesian phylogenetic models,
-#' as created by \link[beautier]{create_inference_mode}.
+#' as created by \link[beautier]{create_inference_model}.
 #'
 #' @seealso Use \link{can_run_mcbette} to see if 'mcbette' can run.
 #' @examples
