+++ man/est_marg_lik.Rd
@@ -7,7 +7,7 @@
 est_marg_lik(
   fasta_filename,
   inference_model = beautier::create_ns_inference_model(),
-  beast2_options = create_mcbette_beast2_options(),
+  beast2_options = beastier::create_mcbette_beast2_options(),
   os = rappdirs::app_dir()$os
 )
 }
@@ -18,7 +18,7 @@ est_marg_lik(
 as can be created by \link[beautier]{create_inference_model}}
 
 \item{beast2_options}{a \code{beast2_options} structure,
-as can be created by \link{create_mcbette_beast2_options}.}
+as can be created by \link[beastier]{create_mcbette_beast2_options}.}
 
 \item{os}{name of the operating system,
 must be \code{unix} (Linux, Mac) or \code{win} (Windows)}
@@ -48,7 +48,7 @@ if (can_run_mcbette()) {
   inference_model <- beautier::create_test_ns_inference_model()
 
   # Setup the options for BEAST2 to be able to call BEAST2 packages
-  beast2_options <- create_mcbette_beast2_options()
+  beast2_options <- beastier::create_mcbette_beast2_options()
 
   # Estimate the marginal likelihood
   marg_lik <- est_marg_lik(
