### 2020-09-05 5e37870
Use regular (i.e. non-test) function, then modify it to a shorter run-time, progress https://github.com/richelbilderbeek/mcbette/issues/30
