+++ vignettes/demo.Rmd
@@ -108,14 +108,22 @@ only differing in their nucleotide substitution model:
 
 ```{r}
 if (can_run_mcbette()) {
-  inference_model_1 <- beautier::create_test_ns_inference_model(
+  # Create the two inference models
+  inference_model_1 <- beautier::create_ns_inference_model(
     site_model = beautier::create_jc69_site_model()
   )
-  inference_model_2 <- beautier::create_test_ns_inference_model(
+  inference_model_2 <- beautier::create_ns_inference_model(
     site_model = beautier::create_gtr_site_model()
   )
+  # Shorten the run, by doing a short (dirty, unreliable) MCMC
+  inference_model_1$mcmc <- beautier::create_test_ns_mcmc()
+  inference_model_2$mcmc <- beautier::create_test_ns_mcmc()
+  
+  
+  # Combine the two inference models
   inference_models <- c(list(inference_model_1), list(inference_model_2))
 
+  # Compare the the two inference models
   df <- est_marg_liks(
     fasta_filename = fasta_filename,
     inference_models = inference_models
