+++ man/est_marg_liks.Rd
@@ -62,12 +62,17 @@ if (can_run_mcbette()) {
   fasta_filename <- system.file("extdata", "simple.fas", package = "mcbette")
 
   # Create two inference models
-  inference_model_1 <- beautier::create_test_ns_inference_model(
+  inference_model_1 <- beautier::create_ns_inference_model(
     site_model = beautier::create_jc69_site_model()
   )
-  inference_model_2 <- beautier::create_test_ns_inference_model(
+  inference_model_2 <- beautier::create_ns_inference_model(
     site_model = beautier::create_hky_site_model()
   )
+
+  # Shorten the run, by doing a short (dirty, unreliable) MCMC
+  inference_model_1$mcmc <- beautier::create_test_ns_mcmc()
+  inference_model_2$mcmc <- beautier::create_test_ns_mcmc()
+
   inference_models <- list(inference_model_1, inference_model_2)
 
   beast2_options <- beastier::create_mcbette_beast2_options()
