### 2020-09-05 richelbilderbeek/mcbette#33 @richelbilderbeek
Add `interpret_bayes_factor_log_lik`
> it would be useful to have an example in the `interpret_bayes_factor` doc of how to use that function with the output from `est_marg_liks`.
