### 2020-09-05 c5b3f05
Process feedback #30, progress https://github.com/ropensci/software-review/issues/360
