### 2020-09-05 47f92b7
Do not test for style in testthat tests, progress #30
