+++ tests/testthat/test-style.R
@@ -1,3 +1,0 @@
-test_that("Package style", {
-  lintr::expect_lint_free()
-})
