### 2020-09-05 311038d
Add 'interpret_marg_lik_estimates', progress #30
