+++ man/default_params_doc.Rd
@@ -17,6 +17,7 @@ default_params_doc(
   fasta_filename,
   inference_model,
   inference_models,
+  marg_liks,
   mcbette_state,
   mcmc,
   os,
@@ -60,6 +61,9 @@ as can be created by \link[beautier]{create_inference_model}}
 
 \item{inference_models}{a list of one or more inference models,
 as can be created by \link[beautier]{create_inference_model}}
+
+\item{marg_liks}{the estimated marginal likelihoods (aka evidences),
+as created by \link{est_marg_liks}}
 
 \item{mcbette_state}{the \link{mcbette} state,
 which is a \link{list} with the following elements:
