+++ NAMESPACE
@@ -8,6 +8,7 @@ export(est_marg_lik)
 export(est_marg_liks)
 export(get_mcbette_state)
 export(interpret_bayes_factor)
+export(interpret_marg_lik_estimates)
 export(mcbette_report)
 export(mcbette_self_test)
 export(set_mcbette_state)
