+++ vignettes/demo.Rmd
@@ -124,23 +124,18 @@ if (can_run_mcbette()) {
   inference_models <- c(list(inference_model_1), list(inference_model_2))
 
   # Compare the the two inference models
-  df <- est_marg_liks(
+  marg_liks <- est_marg_liks(
     fasta_filename = fasta_filename,
     inference_models = inference_models
   )
-  knitr::kable(df)
+  knitr::kable(marg_liks)
 }
 ```
 
-The Bayes factor (usually denoted as K) can be used for model comparison. 
-It can be calculated from the marginal likelihoods ratio:
+The results are interpreted by `interpret_marg_lik_estimates` as follows:
 
 ```{r}
 if (can_run_mcbette()) {
-  marg_lik_jc <- exp(Rmpfr::mpfr(df$marg_log_lik[1], 512))
-  marg_lik_gtr <- exp(Rmpfr::mpfr(df$marg_log_lik[2], 512))
-  bayes_factor <- marg_lik_jc / marg_lik_gtr
-  interpretation <- interpret_bayes_factor(as.numeric(bayes_factor))
-  cat(paste("Interpretation from JC69 model's point of view: ", interpretation))
+  interpret_marg_lik_estimates(marg_liks)
 }
 ```
