### 2020-09-05 6d44f8a
Use 'interpret_marg_lik_estimates', progress #30
