### 2020-09-05 richelbilderbeek/mcbette#33 @richelbilderbeek
I feel this has become redundant after adding `interpret_marg_liks_estimates` (#34). Closing this one :+1:
