### 2020-09-05 3648bd7
Added DOIs, progress #30
