+++ paper.bib
@@ -1,10 +1,13 @@
 @article{Bilderbeek:2018,
   title={babette: BEAUti 2, BEAST2 and Tracer for R},
-  author={Bilderbeek, Richel JC and Etienne, Rampal S},
-  journal={bioRxiv},
-  pages={271866},
+  author={Bilderbeek, Richèl JC and Etienne, Rampal S},
+  journal={Methods in Ecology and Evolution},
+  volume={9},
+  number={9},
+  pages={2034--2040},
   year={2018},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Wiley Online Library},
+  DOI={https://doi.org/10.1111/2041-210X.13032}
 }
 
 @article{Bouckaert:2019,
@@ -15,7 +18,8 @@
   number={4},
   pages={e1006650},
   year={2019},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  DOI={https://doi.org/10.1371/journal.pcbi.1006650}
 }
 
 @article{Drummond:2005,
@@ -26,7 +30,8 @@
   number={5},
   pages={1185--1192},
   year={2005},
-  publisher={Society for Molecular Biology and Evolution}
+  publisher={Society for Molecular Biology and Evolution},
+  DOI={https://doi.org/10.1093/molbev/msi103}
 }
 
 @book{Drummond:2015,
@@ -44,7 +49,8 @@
   number={1309},
   pages={305--311},
   year={1994},
-  publisher={The Royal Society}
+  publisher={The Royal Society},
+  DOI={https://doi.org/10.1098/rstb.1994.0068}
 }
 
 @article{Russel:2019,
@@ -55,6 +61,7 @@
   number={2},
   pages={219--233},
   year={2019},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  DOI={https://doi.org/10.1093/sysbio/syy050}
 }
 
