### 2020-09-05 #360 @richelbilderbeek
I have processed all the feedback, see my feedback here: [https://github.com/richelbilderbeek/mcbette/issues/30](https://github.com/richelbilderbeek/mcbette/issues/30).

There is a (reported) problem with the Travis CI service, hence the build does not pass there. It does pass locally.
