### 2020-09-08 #360 @richelbilderbeek
The first reviewer's comments:

> #### Installation
>
> I had trouble with the package installation due to issues with `rJava`, and never got it working on my Mac (macOS 10.14.6). The CRAN version of `rJava` is looking for a specific JDK which isn't the one on my machine. Compiling it from source appeared to work, but running any function crashes the R console. Installing on a Linux cluster worked.
>
> It looks to me as if the dependency on `rJava` is only used in the `get_default_java_path` function, which seems disproportionate with the amount of trouble it is. I would recommend exploring whether the dependency can be removed, and if not providing a work-around for people who cannot get `rJava` working - one option would be to ask users to provide the path to Java manually.
>

I completely agree that getting `rJava` to work is the toughest nut to crack! I've seen this is true for Linux, Mac and Windows. Yet, `beastier` (a dependency of `mcbette`) needs it to -indeed- get the default Java path as well getting the Java version, which helps out in processing bug reports (created by `beastier::beastier_report`).

However, I feel hand-coding the same subset of functionality is not the way to go. `rJava` is *the* package to work with Java and I hope (and predict) the `rJava` maintainers to simplify its installation. I feel the responsibility about get Java info is at `rJava`, instead of `mcbette`: `mcbette` is there for model comparison.

Note that the `rJava` maintainer strongly recommends upgrading to `rJava 0.9-12` at [the 2020/03/23 rJava release update](https://rforge.net/rJava): maybe that update fixed exactly the problem Joelle had.

> Other comments
>
> instructions on how to configure to use an existing BEAST2 installation would be helpful, if it is possible.

I have documented using BEAST2 installed at a custom location [with this commit](https://github.com/richelbilderbeek/mcbette/commit/bd5b2585837fc6d534983241cfb2f47dd0d370cf).

> On Mac `mcbette` is not looking for it in the default place it is installed to, so that could lead to many duplicate installations.

This would indeed be weird! I looked into this, by writing a test that installs `BEAST2` at a custom location, after which `mcbette` is called to use that BEAST2 install ([see this commit](https://github.com/richelbilderbeek/mcbette/commit/da8c22fa5e79a1769b01899799f41e783379d57e)). The test, however, passes. If this weird behavior could be reproduced, that would definitely be hulpful, but up until then, I cannot fix a problem I cannot reproduce.

> #### Documentation
>
> the documentation (paper, README, vignette) was generally unclear as to which models can be selected using NS - the only mention I found was in the DESCRIPTION, saying that some substitution and clock models are available. It would be useful to include this in the README along with a list of available models. One example showing clock model comparison, in the vignette for instance, would also be useful.

I agree info on this was thin. I've created a vignette in `beautier` that showcases the inference model options.
In README, vignette and paper, I added the available site, clock and tree models.

> in general, the documentation assumes that people using `mcbette` will already be familiar with the other packages of the suite, particularly `beautier`. I think this expectation needs to be stated much more clearly in the README and vignette, and that a link to an appropriate introduction vignette should be provided.

I've described the relation to the `beautier` package in more places and link to a vignette in `beautier` that showcases the inference model options.

> the example in the README should describe the tested model briefly (JC69, strict clock, etc).

I added this.

> there is a `create_ns_inference_model` function in the package, but the examples (README, vignette, package doc) use the `beautier::create_test_ns_inference_model` function instead - it's unclear what the difference is between the two, or which one should be used for a real analysis. If the test function is only used to set up a shorter analysis (as the README seems to imply ?), then it would be better to set the shorter analysis with the real function and modified parameters, and explain what is being done.

I moved `create_ns_inference_model` to the `beautier` package and made the documentation of the `create_[something]_inference_model` functions link to each other.

Additionally, I followed your suggestion and replaced:

```r
x <- create_ns_test_inference_model()
```

by

```r
x <- create_ns_inference_model()
x$mcmc <- create_test_ns_mcmc()
```

> in the default parameter documentation, `beast2_options` should link to the `create_mcbette_beast2_options` function used in all the examples, instead of the  `beastier::create_beast2_options` function. Similarly `inference_model` should link to either `create_ns_inference_model` or `beautier::create_test_ns_inference_model` rather than `beautier::create_inference_model`.

I made these link to one another. Note that the `inference_model` functions have been moves to `beautier`, and the `beast2_options` functions to `beastier`

> when autocompleting in RStudio, the description shown on the function is the beginning of the Description section. This means that for instance `check_mcbette_state` shows only `Will stop otherwise.`. Same issue with `check_beast2_ns_pkg`, `interpret_bayes_factor`.

This is a neat observation, that I was unaware of. I have fixed this.

> contrary to what the doc says, `est_marg_lik` does not return either `estimates` or `trees`.

Well spotted! I've updated the doc in [this commit](https://github.com/richelbilderbeek/mcbette/commit/c2e8f56effe9164ab3885941a0470a93e3b17cf6).

> the error message when `est_marg_lik` fails prints the inference model nicely, but not the BEAST2 options (it shows `beast2_1c6e311fc10d3.xml.stateNANAFALSETRUE`).

Hmmm, I wish I could reproduce that! Because if I look into the code, I do not treat the inference model different than the BEAST2 options:

```r
  tryCatch({
      bbt_run_out <- babette::bbt_run_from_model(
        fasta_filename = fasta_filename,
        inference_model = inference_model,
        beast2_options = beast2_options
      )
      ns <- bbt_run_out$ns
    },
    error = function(e) {
      stop(
        "Could not estimate the marginal likelihood. \n",
        "Error message: ", e$message, "\n",
        "fasta_filename: ", fasta_filename, "\n",
        "beast2_options: ", beast2_options, "\n",
        "inference_model: ", inference_model, "\n"
      )
    }
```

I would enjoy fixing this, but if I cannot check my results, I cannot see if I have fixed it. Could you reproduce the bug? If yes, I will happily add it!

> it would be useful to have an example in the `interpret_bayes_factor` doc of how to use that function with the output from `est_marg_liks`.

I agreed, up until I added `interpret_marg_lik_estimates`, which interprets the results of `est_marg_liks` ([https://github.com/richelbilderbeek/mcbette/issues/34](https://github.com/richelbilderbeek/mcbette/issues/34)), which -IMHO- makes the `interpret_bayes_factor` redundant. I hope you agree!

## Minor comments

> the doc for `est_marg_lik` links to itself, same with `est_marg_liks`

Well spotted! It now links correctly to `est_marg_liks`

>     * `est_marg_lik.R`, l16 "Actual" -> "Current"

Renamed.

### Functionality

> the main feature I think is missing from the package at the moment is the ability to keep the log and tree output of the `babette` runs, in addition to the marginal likelihood. The documentation says that the trees are returned, but that does not appear to be the case. I would expect that users doing model selection will also be interested in the parameter estimates for the best model.

To the `mcbette` main documentation (`?mcbette`) and `mcbette::est_marg_liks` I've added how to obtain the parameter estimates, posterior trees and other things stored in temporary files.

> my other suggestion is to have a function similar to `interpret_bayes_factor` which could directly process the output from `est_marg_liks`, and print the table with the best model highlighted and how all the others compare to that one.

Great idea, I did so by adding `interpret_bayes_factor_log_lik` ([https://github.com/richelbilderbeek/mcbette/issues/33](https://github.com/richelbilderbeek/mcbette/issues/33)). It even comes with a text plot!

> #### JOSS paper comments
>
>     * two references are missing DOIs (required according to JOSS guidelines)

I added DOIs to all journal articles.

>     * there is a `> Example text below` left in l18

Blimey, well spotted! I've put the Summary there :+1:

>     * the single quotes in the title (`'babette'`) give me a yaml parser error when building the file

I removed these quotes altogether. I've added a Travis CI test to verify that `knitr::knit` does work on the file.

>     * l59-61 I think you need to clarify that you are only talking about living species here rather than all species. It's also unclear to me why you say we "should be able to reconstruct" rather than "can reconstruct".

Agreed, I've used to wording you suggested.

>     * l61-62 I would introduce phylogenies earlier, when showing the primates tree.

I rewrote most of the paper, after which phylogenies are introduced sooner.

>     * l70-71 I think this sentence is too vague. I would suggest replacing it with "However, constructing a phylogeny is non-trivial, as it requires the researcher to choose how to model the evolutionary process, including..."

Agreed, I've used your more precise wording. Thanks!

>     * l80-85 This section needs to mention that the 'best' model is always defined in regards to a particular dataset. It's explained later but needs to be made clear earlier.

I have reworded in a such a way to avoid using the word 'best', and it includes that it is dependent on the alignment/dataset:

> `mcbette` is an R package that determines the model with the most
> evidence (aka marginal likelihood) for having generated the alignment,
> from a set of models.

>     * l85 I don't think you can reuse this definition of 'best' for phylogenies - what would it mean for a phylogeny to be "simple" ?

Agreed, I removed it. Thanks!

> Minor comments
>
>     * l49 then -> than

Fixed. Thanks!

>     * l63 a the -> the

Fixed. Thanks!

 * [x] Add `bjoelle` to DESCRIPTION as reviewer

Did so.
