### 2020-09-08 #360 @richelbilderbeek
Second reviewer's reply:

> #### Documentation
>
> The package includes all the following forms of documentation:
>
>  [ ]  **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
> A statement is present, but I do not think it states the problem clearly enough. What is a "Model Comparison"? The phrase is too generic to give a user a sense of what mcbette intends to solve. Although someone familiar with BEAST can likely guess this for him/herself, it should not be up to the reader to interpret. I don't think the author needs to go into a whole lot of detail, but I'd suggest adding 2-3 more sentences to elaborate on what the point of fitting the models is and why the models are being compared at the top of the readme where the package is introduced. There's a few sentences in the `paper.md` that achieve this nicely, so I suggest just incorporating them here as well. I'd also rephrase "solve a common problem" in the Example section to give an explanation of what the problem in the example is and why it is being solved. Again, I anticipate that a user of this package will likely know how all this works. Rather, the point here is to make it more immediately clear what mcbette does when a user first encounters the package.

Indeed, it was a bit concise. I have rephrased this as such:

`mcbette` allows for doing a Model Comparison using `babette` (hence the name),
that is, given a (say, DNA) alignment, it compares multiple phylogenetic inference
models to find the model that is likeliest to generate that alignment.
With this, one can find the phylogenetic inference model that is
simple enough, but not too simple.

To do so, `mcbette` uses `babette` [Bilderbeek and Etienne, 2018] with the addition
of using the BEAST2 [Bouckaert et al., 2019] nested sampling package
as described in [Russell et al., 2019].

>  [x]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
>  In contrast to the other reviewer, I did not encounter any issues with installation or running code (macOS 10.15.4).

Great!

>  [x]  **Vignette(s)** demonstrating major functionality that runs successfully locally
>  Great work! Nice explanation of a clear example. I'll suggest linking this on the readme too as it is very helpful for understanding what the package does and how it works.

Great!

> [x]  **Function Documentation:** for all exported functions in R help

Happy to read  I did a good job :+1:

> [x]  **Examples** for all exported functions in R Help that run successfully locally

Happy to read  I did a good job :+1:

> [x]  **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

Happy to read  I did a good job :+1:

##### For packages co-submitting to JOSS

> * [x]  The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
> The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:

> [ ]  **A short summary** describing the high-level functionality of the software

Indeed, somehow this short summary is missing! I've added it :+1:

> Some of this may be a matter of style, but I have a bit of feedback about how this paper is written. Firstly, the introductory material (all paragraphs and figures before "Constructing a phylogeny, however, is non-trivial...") are too basic for my taste -- I don't think evolution itself needs to be described in fundamental detail at all. This explanation is not appropriate for the package's audience -- we know all this already. Rather, I suggest starting with the logic of paragraph 4 ("Constructing a phylogeny, however, is non-trivial...") which states the problem at hand nicely. Although I like the core idea of this paragraph, the author may also consider fleshing out an explanation of why statistical models are used at all.

Thanks to your feedback, I have removed the whole -indeed slow- introduction and started at the point you suggested. With some rewriting, I feel I have introduced the need for statistical models enough (by introducing the term 'evidence', aka marginal likelihood) and I hope you agree.

> What I will also more strongly recommend is that the author give the reader a sense of where to start when using the package. Broadly speaking, what is the analytical pipeline one should use with mcbette? I think inclusion of a worked example in the paper would help users know where to start, how to structure their data, which functions to look into, and which vignettes they may want to look into. This would be great to have as either the penultimate or the final paragraph. I don't think it needs to go into specific detail, but just give the reader a general sense of how the package works.

Agreed! I have added a section 'Getting Started' that should introduce the reader to the field and direct him/her to the literature.

> * [x]  **Authors:**  A list of authors with their affiliations
> * [x]  **A statement of need** clearly stating problems the software is designed to solve and its target audience.
> * [ ]  **References:** with DOIs for all those that have one (e.g. papers, datasets, software).
>  As noted by the other reviewer, some DOIs are missing and it would be great to have them included

Agreed, thanks! I have added these!

#### Functionality

> [x]  **Installation:** Installation succeeds as documented.
> Nice touch with the `can_run_mcbette()` function! Very handy!

Thanks! If you like that one, you'll love `mcbette_report` :sunglasses: :+1:

> * [x]  **Functionality:** Any functional claims of the software been confirmed.

Cool!

> * [x]  **Performance:** Any performance claims of the software been confirmed.

Awesome.

> * [ ]  **Automated tests:** Unit tests cover essential functions of the package
> and a reasonable range of inputs and conditions. All tests pass on the local machine.
>
> I am unsure why, but testing failed on my machine. Here are the relevant lines
> from the `devtools` check:
>
> > ── Test failures ───────────────────────────────────────────────── testthat ────
> > > library(testthat)
> > > library(mcbette)
> > > before <- get_mcbette_state()
> > > test_check("mcbette")
> > > ── 1. Error: Package style (@test-style.R#2)  ───────────────────────────
> > > invalid 'path' argument
> > > Backtrace:
> >
> >
> >
> > 1. lintr::expect_lint_free()
> > 2. lintr::lint_package(...)
> > 3. base::normalizePath(path, mustWork = FALSE)
> > 4. base::path.expand(path)
> >
> > ══ testthat results  ════════════════════════════════════════════════════
> > [ OK: 88 | SKIPPED: 5 | WARNINGS: 0 | FAILED: 1 ]
> >
> > 1. Error: Package style (@test-style.R#2)
> >
> > Error: testthat unit tests failed
> > Execution halted
> > 1 error x | 0 warnings ✓ | 1 note x
> > Error: R CMD check found ERRORs
> > Execution halted
> > Exited with status 1.

I have the same error. This was not the case upon the submission of `mcbette`.
A bug report has already been filed [here](https://github.com/jimhester/lintr/issues/421) in October 2019.
I've removed the tests, as it is tested by Travis CI anyways.

#### Final approval (post-review)

> * [ ]  **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**

Let's see what this reviewer thinks about this :+1:

> Estimated hours spent reviewing: ~3.5
>
>     * [x]  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

Done, including the ORCID ID :+1:

> ### Review Comments
>
> Firstly, I am quite happy to see that this package exists and was submitted to rOpenSci. `mcbette` is poised to be very helpful towards biologists who work on tree inference and phylogenetic comparative analyses. By using BEAST and accompanying software, `mcbette` builds off the author's previous `babette` and helps expand a user's ability to rely on reproducible pipelines for analyzing genetic data. As someone who will likely use `babette` and other packages in this ecosystem in the near future, I appreciate what `mcbette` brings to the table: giving a reproducible approach to model comparisons of sequence evolution.
>
> In running the worked examples as well as feeding in sequences from my own data sets, I did not encounter any issues with how the code ran or the results it produced. To the best of my ability to assess this, `mcbette` seems to produce correct results.

Happy to hear this!

> That being said, testing failed on my machine. Please see above for the relevant lines from the `devtools` check output. Also in investigating `testthat.R`, the result of line 6 `test_check("mcbette")` was:
>
> > Error: No tests found for mcbette

As noted above, the devtools issues are caused by something upstream (in the lintr package) and needs to be resolved there.

Regarding the second error, I can perfectly reproduce this. I predict this is a feature, not a bug. With CTRL+SHIFT+T in R Studio or `devtools::test()` the tests do start as expected. Also, mcbette is tested by Travis CI (for Linux and Mac) and AppVeyor (for Windows) and its code coverage is measured from these tests. Therefore, I assume the tests works and that the `No tests found for mcbette` message is some R feature I and you misinterpret.

> Unlike the other reviewer, I did not encounter issues during install (using macOS 10.15.4 with R 3.6.2 and then again on R 4.0.0) so I am sorry to not be able to offer any insight on troubleshooting. I should note that I already had `rJava` installed, which may help explain the lack of problems.

I am happy to have this confirmed :+1:

> My inline comments above are geared towards enhancing the clarity of the documentation. I will agree with the other reviewer's assessment -- the documentation and the accompanying paper are a little too sparse as presently written. I'd appreciate seeing expansions of both the high-level functionality (in a broad sense) as well as more detailed descriptions of how the author envisions users feeding data into analytical pipelines. I don't mind having similar information repeated in the readme, description, vignettes, and paper; doing so helps not only reinforce the message of the package and also helps users' understanding regardless of their point of entry into the package.

I added quite some documentation, due to feedback from the other reviewer indeed. I hope this will also satisfy your needs.

> I agree with all the specific bullet points the other reviewer raised regarding the Documentation and Functionality sections and for the sake of brevity won't repeat them here. I'll add that providing a full(er) list of potential models would be great to have, perhaps as its own vignette.
>
> I do not necessarily agree with the other reviewer's comments regarding the JOSS paper but this is primarily because I feel the paper should be restructured (please see inline comments above).

I have restructured the paper from a mix of both reviews. I hope you both like it. I, for sure, do!

> Thanks again for the opportunity to review this package. Again, my suggestions are motivated by wanting to enhance the clarity of presentation and make it easier for potential users to see the value of relying on `mcbette`. It's nice work!
>
> turtle

Thanks so much for reviewing!
