### 2020-09-08 richelbilderbeek/mcbette#30 @richelbilderbeek
Posted these replies back on [https://github.com/ropensci/software-review/issues/360](https://github.com/ropensci/software-review/issues/360).

Closing this one :+1:
