### 2020-09-09 #360 @vbaliga
Rather than edit my previous review, I am electing to write a fresh one here:

  Again, I had no trouble with installation but note that I already use rjava for other things

>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>
>- [x] **Authors:**  A list of authors with their affiliations
>
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>
> Very nicely done!
>
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).
>
As noted by the other reviewer, some DOIs are missing and it would be great to have them included

Estimated hours spent reviewing: <1 for this round

---

Thanks very much for taking my feedback into account. This package was already very good and I think it is even better now in the sense that it is more approachable for new users.

I have a couple minor points to bring up, but please note that despite these points I am satisfied with the changes I already see and I am supportive of this work's publication.

During devtools::check(), I encountered:
``` r
> checking R code for possible problems ... NOTE
  interpret_marg_lik_estimates: no visible global function definition for
    ‘na.omit’
    na.omit
  Consider adding
    importFrom("stats", "na.omit")
  to your NAMESPACE file.
```

The other point is that I think it would be a nice touch to link to the vignettes at the end of `paper.md`. The paper does a nice job giving an overview of the package and it would be great to end it by giving users a suggestion of what they might want to read next, i.e. the vignettes.

Thanks again to @maelle for the opportunity to review this and to @richelbilderbeek for taking my feedback. I am looking forward to using `mcbette` and Richel's other packages in the near future.

Vikram

🐢
