### 2020-09-10 d835cc5
Mention demo vignette, thanks @vbaliga, progress https://github.com/ropensci/software-review/issues/360
