### 2020-09-10 dbb308e
Add namespace to 'na.omit' call, thanks @vbaliga, progress https://github.com/ropensci/software-review/issues/360
