+++ scripts/run.R
@@ -22,7 +22,7 @@ if (can_run_mcbette()) {
   # Create an ordered data frame
   ################################################################################
   # Keep rows without an NA
-  df_ordered <- na.omit(df)
+  df_ordered <- stats::na.omit(df)
 
   # Order from high to low
   df_ordered <- df_ordered[ order(-df_ordered$weight), ]
