+++ .travis.yml
@@ -101,3 +101,4 @@ after_failure:
   # Download thesis
   - Rscript -e 'download.file("https://www.rug.nl/research/portal/files/132028374/Complete_thesis.pdf", tempfile())'
 
+
