### 2020-09-22 f290f51
Do not install ImageMagick from PPA
