+++ .travis.yml
@@ -42,7 +42,7 @@ before_install:
   # From https://stackoverflow.com/a/20042195
   - Rscript -e "update.packages(checkBuilt = TRUE, ask = FALSE)"
   # ImageMagick
-  - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo add-apt-repository -y ppa:opencpu/imagemagick; fi
+  # - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo add-apt-repository -y ppa:opencpu/imagemagick; fi
   - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo apt-get update; fi
   - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo apt-get install -y libmagick++-dev; fi
   # Rmpfr
