### 2020-09-23 fb28af7
No OSX, due to 'sudo tlmgr update --self' taking too long, see https://travis-ci.org/github/richelbilderbeek/mcbette/jobs/729580427#L109
