### 2020-09-24 31c51d6
No testthat tests in example code
