+++ R/est_marg_lik.R
@@ -15,8 +15,6 @@
 #' @examples
 #' if (can_run_mcbette()) {
 #'
-#'   library(testthat)
-#'
 #'   # An example FASTA file
 #'   fasta_filename <- system.file("extdata", "simple.fas", package = "mcbette")
 #'
