### 2020-09-29 #360 @bjoelle
Sorry it took me so long to come back to this review ! Overall I am happy with the changes that have been made to the package and documentation.
My one comment would be that I feel like the current version of the JOSS paper underplays the usefulness of the package (the statement of need in particular feels a bit weak). I would state more clearly what the added value of `mcbette` is (compared to using BEAST2 directly) : easy set-up and execution of several parallel analyses using different models, easy analysis and nice printing of the results of the model selection, possibility to integrate it into an existing R pipeline, etc.
