### 2020-10-08 ropensci/mcbette#36 @richelbilderbeek
Process last rOpenSci feedback
See feedback [here](https://github.com/ropensci/software-review/issues/360)
