### 2020-10-09 d530fb0
Add CONTRIBUTING.md and Code of Conduct
