### 2020-10-09 4bddc64
Processed feedback bjoelle, thank bjoelle, progress https://github.com/ropensci/software-review/issues/360
