### 2020-10-09 4820aa8
Use rOpenSci Code of Conduct, progress #36
