+++ DESCRIPTION
@@ -32,8 +32,8 @@ License: GPL-3
 LazyData: true
 RoxygenNote: 7.1.1
 VignetteBuilder: knitr
-URL: https://github.com/richelbilderbeek/mcbette
-BugReports: https://github.com/richelbilderbeek/mcbette/issues
+URL: https://github.com/ropensci/mcbette
+BugReports: https://github.com/ropensci/mcbette/issues
 Imports:
     babette,
     beautier (>= 2.4),
