### 2020-10-09 6a6de65
Rebuild 'README.md' using 'devtools::build_readme()'
