### 2020-10-09 9f5e71e
Ran 'codemetar::write_codemeta()', progress #36
