### 2020-10-09 ropensci/mcbette#36 @richelbilderbeek
Done most (details at [here](https://github.com/ropensci/mcbette/issues/36)), will continue tomorrow.
