### 2020-10-13 93fc36e
Add LICENSE file for https://github.com/ropensci/mcbette/issues/37#issuecomment-706587216
