### 2020-10-13 a38c252
Use smaller logo for #37
