+++ paper.bib
@@ -7,7 +7,7 @@
   pages={2034--2040},
   year={2018},
   publisher={Wiley Online Library},
-  DOI={https://doi.org/10.1111/2041-210X.13032}
+  DOI={10.1111/2041-210X.13032}
 }
 
 @article{Bouckaert:2019,
@@ -19,7 +19,7 @@
   pages={e1006650},
   year={2019},
   publisher={Public Library of Science},
-  DOI={https://doi.org/10.1371/journal.pcbi.1006650}
+  DOI={10.1371/journal.pcbi.1006650}
 }
 
 @article{Drummond:2005,
@@ -31,7 +31,7 @@
   pages={1185--1192},
   year={2005},
   publisher={Society for Molecular Biology and Evolution},
-  DOI={https://doi.org/10.1093/molbev/msi103}
+  DOI={10.1093/molbev/msi103}
 }
 
 @book{Drummond:2015,
@@ -50,7 +50,7 @@
   pages={305--311},
   year={1994},
   publisher={The Royal Society},
-  DOI={https://doi.org/10.1098/rstb.1994.0068}
+  DOI={10.1098/rstb.1994.0068}
 }
 
 @article{Russel:2019,
@@ -62,6 +62,6 @@
   pages={219--233},
   year={2019},
   publisher={Oxford University Press},
-  DOI={https://doi.org/10.1093/sysbio/syy050}
+  DOI={10.1093/sysbio/syy050}
 }
 
