### 2020-10-15 dfb8d2d
Use CRAN versions
