+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: mcbette
 Title: Model Comparison Using 'babette'
-Version: 1.9
+Version: 1.10
 Authors@R: c(
     person(
       "Richèl J.C.", "Bilderbeek", 
