### 2020-10-16 90b4047
Add trailing slash to URL, thanks Uwe Ligges
