+++ DESCRIPTION
@@ -32,7 +32,7 @@ License: GPL-3 | file LICENSE
 LazyData: true
 RoxygenNote: 7.1.1
 VignetteBuilder: knitr
-URL: https://github.com/ropensci/mcbette
+URL: https://github.com/ropensci/mcbette/
 BugReports: https://github.com/ropensci/mcbette/issues
 Imports:
     babette,
