+++ .travis.yml
@@ -34,7 +34,6 @@ r_github_packages:
   - jimhester/covr
   - jimhester/lintr
   - MangoTheCat/goodpractice
-  - ropensci/beautier
   #- ropensci/tracerer
   - ropensci/beastier
   #- ropensci/mauricer
