### 2020-10-16 937ac57
Fix 'Invalid license file pointers', https://travis-ci.org/github/ropensci/mcbette/jobs/736062911#L1291
