### 2020-10-16 928cce0
Simplify script
