### 2020-10-16 53b92d6
Use LICENSE file, do not mention it is DESCRIPTION, due to 'Invalid license file pointers: LICENSE' note
