### 2020-10-16 ropensci/mcbette#36 @richelbilderbeek
Done, posted reply at the original post at [https://github.com/ropensci/software-review/issues/360](https://github.com/ropensci/software-review/issues/360)
