### 2020-10-17 e590868
beastier v2.2 is on CRAN
