### 2020-10-26 #360 @richelbilderbeek
Hi @stefaniebutland,

> Do you think this one would be suitable as a tech note, rather than longer narrative blog post?

I would prefer the longer narrative blog post, as this one is for a broader audience. I do wonder if that is not too similar to the JOSS acticle I submitted. What do you think?

 * If you think the JOSS article is too similar to a longer narrative blog post, I will happily write a tech note
 * Else, I will happily write a longer narrative blog post

:+1:
