### 2020-10-27 #360 @stefaniebutland
I'm happy to leave this up to your judgement @richelbilderbeek. The post will reach a different audience, so a good way to get more attention for it. That will work as long as you make the content different enough from the JOSS paper. A blog post allows for a somewhat more casual tone.

You may use 2020-11-17 or 2020-12-01 as a tentative publication date; date to be finalized during review.
