### 2020-11-03 #360 @stefaniebutland
Hi @richelbilderbeek. Do you prefer 2020-11-17 or 2020-12-01 as a tentative publication date? If 2020-11-17, fyi my colleague @steffilazerte will review on 2020-11-13, so a short turnaround time for edits.
